Drasil.Code.CodeExpr

Description

Re-export CodeExpr constructors.

Documentation

data CodeExpr

Expression language where all terms also denote a term in GOOL (i.e. translation is total and meaning preserving).

class CodeExprC r where

Methods

new :: (Callable f, HasUID f, CodeIdea f) => f -> [r] -> r

Constructs a CodeExpr for actor creation (constructor call)

newWithNamedArgs :: (Callable f, HasUID f, CodeIdea f, HasUID a, IsArgumentName a) => f -> [r] -> [(a, r)] -> r

Constructs a CodeExpr for actor creation (constructor call) that uses named arguments

msg :: (Callable f, HasUID f, CodeIdea f, HasUID c, HasSpace c, CodeIdea c) => c -> f -> [r] -> r

Constructs a CodeExpr for actor messaging (method call)

msgWithNamedArgs :: (Callable f, HasUID f, CodeIdea f, HasUID c, HasSpace c, CodeIdea c, HasUID a, IsArgumentName a) => c -> f -> [r] -> [(a, r)] -> r

Constructs a CodeExpr for actor messaging (method call) that uses named arguments

field :: CodeVarChunk -> CodeVarChunk -> r

Constructs a CodeExpr representing the field of an actor

applyWithNamedArgs :: (HasUID f, HasSymbol f, HasUID a, IsArgumentName a) => f -> [r] -> [(a, r)] -> r

Similar to apply, but takes a relation to apply to FCall.

class ExprC r where

Methods

lit :: Literal -> r

($=) :: r -> r -> r infixr 4

($!=) :: r -> r -> r

($<) :: r -> r -> r

Smart constructor for ordering two equations.

($>) :: r -> r -> r

Smart constructor for ordering two equations.

($<=) :: r -> r -> r

Smart constructor for ordering two equations.

($>=) :: r -> r -> r

Smart constructor for ordering two equations.

($.) :: r -> r -> r

Smart constructor for the dot product of two equations.

($+) :: r -> r -> r

Add two expressions.

($*) :: r -> r -> r

Multiply two expressions.

($-) :: r -> r -> r

($/) :: r -> r -> r infixl 7

($^) :: r -> r -> r infixr 8

($=>) :: r -> r -> r

($<=>) :: r -> r -> r

($&&) :: r -> r -> r infixr 9

($||) :: r -> r -> r infixr 9

in' :: r -> r -> r

Smart constructor for set-theoretic membership relation. Added ' to avoid conflict.

abs_ :: r -> r

Smart constructor for taking the absolute value of an expression.

neg :: r -> r

Smart constructor for negating an expression.

log :: r -> r

Smart constructor to take the log of an expression.

ln :: r -> r

Smart constructor to take the ln of an expression.

sqrt :: r -> r

Smart constructor to take the square root of an expression.

sin :: r -> r

Smart constructor to apply sin to an expression.

cos :: r -> r

Smart constructor to apply cos to an expression.

tan :: r -> r

Smart constructor to apply tan to an expression.

sec :: r -> r

Smart constructor to apply sec to an expression.

csc :: r -> r

Smart constructor to apply csc to an expression.

cot :: r -> r

Smart constructor to apply cot to an expression.

arcsin :: r -> r

Smart constructor to apply arcsin to an expression.

arccos :: r -> r

Smart constructor to apply arccos to an expression.

arctan :: r -> r

Smart constructor to apply arctan to an expression.

exp :: r -> r

Smart constructor for the exponential (base e) function.

dim :: r -> r

Smart constructor for calculating the dimension of a vector.

norm :: r -> r

Smart constructor for calculating the normal form of a vector.

negVec :: r -> r

Smart constructor for negating vectors.

not_ :: r -> r

Smart constructor for applying logical negation to an expression.

idx :: r -> r -> r

Smart constructor for indexing.

idxOf :: r -> r -> r

Smart constructor for indexOf. Finds the index of the first occurrence of a value in a list.

defint :: Symbol -> r -> r -> r -> r

Smart constructor for the summation, product, and integral functions over an interval.

defsum :: Symbol -> r -> r -> r -> r

Smart constructor for the summation, product, and integral functions over an interval.

defprod :: Symbol -> r -> r -> r -> r

Smart constructor for the summation, product, and integral functions over an interval.

realInterval :: HasUID c => c -> RealInterval r r -> r

Smart constructor for 'real interval' membership.

euclidean :: [r] -> r

Euclidean function : takes a vector and returns the sqrt of the sum-of-squares.

cross :: r -> r -> r

Smart constructor to cross product two expressions.

vScale :: r -> r -> r

Smart constructor for vector scaling

vAdd :: r -> r -> r

Vector Addition

vSub :: r -> r -> r

Vector Subtraction

completeCase :: [(r, r)] -> r

Smart constructor for case statements with a complete set of cases.

incompleteCase :: [(r, r)] -> r

Smart constructor for case statements with an incomplete set of cases.

matrix :: [[r]] -> r

Create a matrix.

set' :: Space -> [r] -> r

Create a Set.

apply :: (HasUID f, HasSymbol f) => f -> [r] -> r

Applies a given function with a list of parameters.

sy :: (HasUID c, HasSymbol c) => c -> r

Create an Expr from a Symbolic Chunk.

class LiteralC r where

Methods

int :: Integer -> r

str :: String -> r

dbl :: Double -> r

exactDbl :: Integer -> r

perc :: Integer -> Integer -> r

expr :: Expr -> CodeExpr

Render an algebraic expression into our code expression language.