Drasil.Code.CodeExpr.Extract

Documentation

eDep :: CodeExpr -> [UID]

Get dependencies from an equation.

eDep' :: CodeExpr -> [UID]

Get dependencies from an equation, without functions.

eNamesRI :: RealInterval CodeExpr CodeExpr -> [UID]

Generic traversal of everything that could come from an interval to names (similar to eNames).

eNamesRI' :: RealInterval CodeExpr CodeExpr -> [UID]

Generic traversal of everything that could come from an interval to names without functions (similar to eNames').