Drasil.Code.CodeVar

Description

Defines chunk types for use in code generation.

Documentation

class CodeIdea c where

A CodeIdea must include some code and its name.

Methods

codeName :: c -> String

Name of the idea.

codeChunk :: c -> CodeChunk

Code chunk associated with the idea.

class CodeIdea c => DefiningCodeExpr c where

A DefiningCodeExpr must have it's underlying chunk defined in the CodeExpr language.

Methods

codeExpr :: Lens' c CodeExpr

programName :: CommonIdea c => c -> String

Convert an abbreviation into one deemed 'code-friendly', removing spaces, and replacing special characters with underscores.

FIXME: This should NOT be treated as a getter, but something we cache local to something that has a 'program name'.

funcPrefix :: String

Used when a function name needs to be distinguishable from a variable name.

data VarOrFunc

Details if a piece of code is meant to be a variable or a function.

Constructors

Var

Func

data CodeChunk

Basic chunk representation in the code generation context. Contains a DefinedQuantityDict and the kind of code (variable or function).

Constructors

CodeC

Fields

_qc :: DefinedQuantityDict

kind :: VarOrFunc

qc :: Lens' CodeChunk DefinedQuantityDict

data CodeVarChunk

Chunk representing a variable. The obv field represents the object containing this variable, if it is an object field.

Constructors

CodeVC

Fields

_ccv :: CodeChunk

_obv :: Maybe CodeChunk

obv :: Lens' CodeVarChunk (Maybe CodeChunk)

ccv :: Lens' CodeVarChunk CodeChunk

newtype CodeFuncChunk

Chunk representing a function.

Constructors

CodeFC

Fields

_ccf :: CodeChunk

ccf :: Iso' CodeFuncChunk CodeChunk

listToArray :: CodeVarChunk -> CodeVarChunk