Drasil.Database.UID

Description

Unique Identifier used across Drasil.

Documentation

data UID

A UID is a 'unique identifier' for things that we will put into our database of information. We use a newtype wrapper to make sure we are only using UIDs where desired.

class HasUID c where

The most basic item: having a unique identifier key, here a UID.

Methods

uid :: Getter c UID

Provides a unique id for internal Drasil use.

mkUid :: String -> UID

Smart constructor for UIDs from raw Values.

nsUid :: String -> UID -> UID

Nest a UID under a namespace.

(+++) :: HasUID a => a -> String -> UID

Append a suffix to a thing with a UID and get the resulting UID.

(+++.) :: UID -> String -> UID

Append a suffix to a UID.

(+++!) :: (HasUID a, HasUID b) => a -> b -> UID

Merge the UIDs of two chunks through simple concatenation.

showUID :: HasUID a => a -> String

Get something's UID as a Value.