Language.Drasil

Documentation

data Expr where

Drasil Expressions

Constructors

Dbl :: Double -> Expr

Int :: Integer -> Expr

ExactDbl :: Integer -> Expr

Str :: String -> Expr

Perc :: Integer -> Integer -> Expr

AssocA :: AssocArithOper -> [Expr] -> Expr

AssocB :: AssocBoolOper -> [Expr] -> Expr

Deriv :: DerivType -> Expr -> UID -> Expr
Derivative, syntax is: Type (Partial or total) -> principal part of change -> with respect to For example: Deriv Part y x1 would be (dy/dx1)

C :: UID -> Expr
C stands for Chunk, for referring to a chunk in an expression. Implicitly assumes has a symbol.

FCall :: UID -> [Expr] -> [(UID, Expr)] -> Expr
F(x) is (FCall F [x] []) or similar. FCall accepts a list of params and a list of named params. F(x,y) would be (FCall F [x,y]) or sim. F(x,n=y) would be (FCall F [x] [(n,y)]).

New :: UID -> [Expr] -> [(UID, Expr)] -> Expr
Actor creation given UID and parameters

Message :: UID -> UID -> [Expr] -> [(UID, Expr)] -> Expr
Message an actor: 1st UID is the actor, 2nd UID is the method

Field :: UID -> UID -> Expr
Access a field of an actor: 1st UID is the actor, 2nd UID is the field

Case :: Completeness -> [(Expr, Relation)] -> Expr
For multi-case expressions, each pair represents one case

Matrix :: [[Expr]] -> Expr

UnaryOp :: UFunc -> Expr -> Expr
Unary functions/operations

UnaryOpB :: UFuncB -> Expr -> Expr

UnaryOpVec :: UFuncVec -> Expr -> Expr

ArithBinaryOp :: ArithBinOp -> Expr -> Expr -> Expr
Binary functions/operations

BoolBinaryOp :: BoolBinOp -> Expr -> Expr -> Expr

EqBinaryOp :: EqBinOp -> Expr -> Expr -> Expr

LABinaryOp :: LABinOp -> Expr -> Expr -> Expr

OrdBinaryOp :: OrdBinOp -> Expr -> Expr -> Expr

VVVBinaryOp :: VVVBinOp -> Expr -> Expr -> Expr

VVNBinaryOp :: VVNBinOp -> Expr -> Expr -> Expr

Operator :: AssocArithOper -> DomainDesc Expr Expr -> Expr -> Expr
Operators are generalized arithmetic operators over a |DomainDesc| of an |Expr|. Could be called |BigOp|. ex: Summation is represented via |Add| over a discrete domain

IsIn :: Expr -> Space -> Expr
element of

RealI :: UID -> RealInterval Expr Expr -> Expr
a different kind of 'element of'

data UFunc

Unary functions

Constructors

Abs

Log

Ln

Sin

Cos

Tan

Sec

Csc

Cot

Arcsin

Arccos

Arctan

Exp

Sqrt

Neg

data UFuncB

data UFuncVec

data ArithBinOp

data BoolBinOp

data EqBinOp

data LABinOp

data OrdBinOp

data VVVBinOp

data VVNBinOp

data AssocArithOper

Constructors

AddI

AddRe

MulI

MulRe

data AssocBoolOper

Constructors

And

Or

data DerivType

Constructors

Part

Total

data Completeness

For case expressions

Constructors

Complete

Incomplete

type Relation = Expr

($=) :: Expr -> Expr -> Expr infixr 4

($<) :: Expr -> Expr -> Expr

($<=) :: Expr -> Expr -> Expr

($>) :: Expr -> Expr -> Expr

($>=) :: Expr -> Expr -> Expr

($^) :: Expr -> Expr -> Expr infixr 8

($&&) :: Expr -> Expr -> Expr infixr 9

($||) :: Expr -> Expr -> Expr infixr 9

($=>) :: Expr -> Expr -> Expr

($<=>) :: Expr -> Expr -> Expr

($.) :: Expr -> Expr -> Expr

($-) :: Expr -> Expr -> Expr

($/) :: Expr -> Expr -> Expr infixl 7

addI :: Expr -> Expr -> Expr

addRe :: Expr -> Expr -> Expr

mulI :: Expr -> Expr -> Expr

mulRe :: Expr -> Expr -> Expr

dep :: Expr -> [String]

Get dependencies from an equation

abs_ :: Expr -> Expr

Smart constructor for taking the absolute value of an expression

neg :: Expr -> Expr

Smart constructor for negating an expression

log :: Expr -> Expr

Smart constructor to take the log of an expression

ln :: Expr -> Expr

Smart constructor to take the ln of an expression

abs :: Num a => a -> a

Absolute value.

sin :: Expr -> Expr

Smart constructor to apply sin to an expression

cos :: Expr -> Expr

Smart constructor to apply cos to an expression

tan :: Expr -> Expr

Smart constructor to apply tan to an expression

sec :: Expr -> Expr

Smart constructor to apply sec to an expression

csc :: Expr -> Expr

Smart constructor to apply csc to an expression

cot :: Expr -> Expr

Smart constructor to apply cot to an expression

arcsin :: Expr -> Expr

Smart constructor to apply arcsin to an expression

arccos :: Expr -> Expr

Smart constructor to apply arccos to an expression

arctan :: Expr -> Expr

Smart constructor to apply arctan to an expression

exp :: Expr -> Expr

Smart constructor for the exponential (base e) function

sqrt :: Expr -> Expr

Smart constructor to take the square root of an expression

euclidean :: [Expr] -> Expr

Euclidean function : takes a vector and returns the sqrt of the sum-of-squares

norm :: Expr -> Expr

Smart constructor for calculating the normal form of a vector

not_ :: Expr -> Expr

Smart constructor for applying logical negation to an expression

square :: Expr -> Expr

Smart constructor to square a function

half :: Expr -> Expr

oneHalf :: Expr

oneThird :: Expr

recip_ :: Expr -> Expr

Smart constructor for 1/x expressions

dim :: Expr -> Expr

Smart constructor for calculating the dimension of a vector

idx :: Expr -> Expr -> Expr

Smart constructor for indexing

int :: Integer -> Expr

Smart constructor for integers

dbl :: Double -> Expr

Smart constructor for doubles

exactDbl :: Integer -> Expr

Smart constructor for exact doubles

frac :: Integer -> Integer -> Expr

Smart constructor for fractions

str :: String -> Expr

Smart constructor for strings

perc :: Integer -> Integer -> Expr

Smart constructors for percents

isin :: Expr -> Space -> Expr

Smart constructor for set membership FIXME: first argument really ought to be restricted to a variable, as IsIn should only be used as a type proxy

completeCase :: [(Expr, Relation)] -> Expr

Smart constructor for case statement with complete set of cases

incompleteCase :: [(Expr, Relation)] -> Expr

Smart constructor for case statement with incomplete set of cases

sumAll :: Symbol -> Expr -> Expr

defsum :: Symbol -> Expr -> Expr -> Expr -> Expr

Smart constructor for the summation, product, and integrals

prodAll :: Symbol -> Expr -> Expr

defprod :: Symbol -> Expr -> Expr -> Expr -> Expr

Smart constructor for the summation, product, and integrals

defint :: Symbol -> Expr -> Expr -> Expr -> Expr

Smart constructor for the summation, product, and integrals

intAll :: Symbol -> Expr -> Expr

realInterval :: HasUID c => c -> RealInterval Expr Expr -> Expr

Smart constructor for 'real interval' membership

deriv :: (HasUID c, HasSymbol c) => Expr -> c -> Expr

Gets the derivative of an Expr with respect to a Symbol

pderiv :: (HasUID c, HasSymbol c) => Expr -> c -> Expr

Gets the derivative of an Expr with respect to a Symbol

sy :: (HasUID c, HasSymbol c) => c -> Expr

Get an Expr from a Symbol

apply :: (HasUID f, HasSymbol f) => f -> [Expr] -> Expr

apply1 :: (HasUID f, HasSymbol f, HasUID a, HasSymbol a) => f -> a -> Expr

apply2 :: (HasUID f, HasSymbol f, HasUID a, HasSymbol a, HasUID b, HasSymbol b) => f -> a -> b -> Expr

applyWithNamedArgs :: (HasUID f, HasSymbol f, HasUID a, IsArgumentName a) => f -> [Expr] -> [(a, Expr)] -> Expr

cross :: Expr -> Expr -> Expr

Smart constructor to cross product two expressions

m2x2 :: Expr -> Expr -> Expr -> Expr -> Expr

Matrix helper function

vec2D :: Expr -> Expr -> Expr

Matrix helper function

dgnl2x2 :: Expr -> Expr -> Expr

Matrix helper function

data Special

Special characters including partial derivatives, degree circle, and underscores

Constructors

Partial

Circle

class RenderSpecial r where

Class for rendering special characters

Methods

special :: Special -> r

type UID = String

class HasUID c where

The most basic item: having a unique key, here a UID

Methods

uid :: Lens' c UID

Provides a unique id for internal Drasil use

class HasShortName s where

A ShortName is the text to be displayed for a link. Used for referencing within a document that can include symbols and whatnot if required. Visible in the typeset documents (pdf)

Methods

shortname :: s -> ShortName

class HasRefAddress b where

For a "Reference Address", we just need a getter.

Methods

getRefAdd :: b -> String

class HasSymbol c where

A HasSymbol is anything which has a Symbol

Methods

symbol :: c -> Stage -> Symbol

Provides the Symbol for a particular stage of generation

class HasFields c where

Citations have Fields

Methods

getFields :: Lens' c [CiteField]

class HasUID c => NamedIdea c where

A NamedIdea is a term that we've identified (has an id) as being worthy of naming.

Methods

term :: Lens' c NP

Lens to the term (a noun phrase)

class HasAdditionalNotes c where

Methods

getNotes :: Lens' c [Sentence]

class HasSpace c where

HasSpace is anything which has a Space...

Methods

typ :: Lens' c Space

class HasUnitSymbol u where

Some chunks store a unit symbol

Methods

usymb :: u -> USymb

class HasReference c where

A class that contains a list of References

Methods

getReferences :: Lens' c [Reference]

class HasReasVal c where

A HasReasVal is a Quantity that could have a reasonable value

Methods

reasVal :: Lens' c (Maybe Expr)

class HasDerivation c where

A class that might have a Derivation

Methods

derivations :: Lens' c (Maybe Derivation)

class NamedIdea c => Idea c where

An |Idea| is the meet of |NamedIdea| and |CommonIdea|. In other words, it may have an acronym/abbreviation.

Methods

getA :: c -> Maybe String

class Definition c where

Methods

defn :: Lens' c Sentence

defn provides (a Lens to) the definition for a chunk

class ConceptDomain c where

Methods

cdom :: c -> [UID]

cdom provides Getter for the concept domain tags for a chunk

cdom should be exported for use by the Drasil framework, but should not be exported beyond that.

type Concept c = (Idea c, Definition c, ConceptDomain c)

Concepts are Ideas with definitions and domains

class (Idea u, Definition u, HasUnitSymbol u) => IsUnit u where

Units are Ideas with a Definition which store a unit symbol. They must also be explicitly declared to be instances of IsUnit

Minimal complete definition

udefn, getUnits

Methods

getUnits :: u -> [UID]

class NamedIdea c => CommonIdea c where

CommonIdea is a NamedIdea with the additional constraint that it must have an abbreviation.

Methods

abrv :: c -> String

Introduces abrv which necessarily provides an abbreviation.

class Constrained c where

A Constrained is a Quantity that has value constraints but do not enforce Quantity at this point

Methods

constraints :: Lens' c [Constraint]

class ExprRelat c where

Has a function that turns into a Relation

Methods

relat :: c -> Relation

class DefiningExpr c where

A better version of ExprRelat

Methods

defnExpr :: Lens' c Expr

class HasUncertainty c where

HasUncertainty is just a chunk with some uncertainty associated to it. This uncertainty is represented as a decimal value between 0 and 1 (percentage).

Methods

unc :: Lens' c Uncertainty

class (Idea c, HasSpace c, HasSymbol c) => Quantity c

A Quantity is an Idea with a Space and a symbol. In theory, it should also have MayHaveUnit, but that causes all sorts of import cycles (or lots of orphans)

class HasSymbol c => Callable c

Some chunks can be called like functions

class HasSymbol c => IsArgumentName c

cw :: Concept c => c -> ConceptChunk

For projecting out to the ConceptChunk data-type.

data ConceptChunk

The ConceptChunk datatype is a Concept that contains an idea (IdeaDict), a definition (Sentence), and a domain ([UID]).

data CommonConcept

Contains a common idea (CI) with a definition (Sentence).

data ConceptInstance

Contains a ConceptChunk, reference address, and a ShortName.

sDom :: [UID] -> UID

Check if something has one domain. Throws an error if there is more than one.

commonIdea :: String -> NP -> String -> [UID] -> CI

The commonIdea smart constructor requires a chunk id (UID), a term (NP), an abbreviation (String), and a domain ([UID]).

data CI

The common idea (with NounPhrase) data type. It must have a UID, NounPhrase for its term, an abbreviation (String), and a domain ([UID]).

getAcc :: CI -> Sentence

Get abbreviation in Sentence form from a CI.

getAccStr :: CI -> String

Get abbreviation in String form from a CI.

commonIdeaWithDict :: String -> NP -> String -> [IdeaDict] -> CI

Similar to commonIdea, but takes a list of IdeaDict (often a domain).

prependAbrv :: CommonIdea c => c -> String -> String

Prepends the abbreviation from a CommonIdea to a String.

data NamedArgument

Any quantity can be a named argument (wrapper for QuantityDict)

narg :: (Quantity q, MayHaveUnit q) => q -> NamedArgument

Smart constructor for NamedArgument

data NamedChunk

Note that a NamedChunk does not have an acronym/abbreviation as that's a CommonIdea, which has its own representation. Contains a UID and a term (NP).

short :: (Idea c, HasUID c) => c -> Sentence

Get short form (if it exists), else get term.

nc :: String -> NP -> NamedChunk

NamedChunk constructor, takes a UID and a term.

data IdeaDict

IdeaDict is the canonical dictionary associated to Idea. Contains a NamedChunk and maybe an abbreviation (String). Don't export the record accessors.

mkIdea :: String -> NP -> Maybe String -> IdeaDict

IdeaDict constructor, takes a UID, NP, and an abbreviation in the form of Maybe String

nw :: Idea c => c -> IdeaDict

Historical name: nw comes from 'named wrapped' from when NamedIdea exported getA (now in Idea). But there are no more wrappers, instead we have explicit dictionaries. Unwraps an Idea and places its UID and NP into an IdeaDict with Nothing for an abbreviation.

physc :: RealInterval Expr Expr -> Constraint

Range of Physical constraints

sfwrc :: RealInterval Expr Expr -> Constraint

Range of Software constraints

enumc :: [Double] -> Constraint

Enumeration of values, default to Physical

isPhysC :: Constraint -> Bool

Helpful for filtering for Physical / Software constraints

isSfwrC :: Constraint -> Bool

Helpful for filtering for Physical / Software constraints

data Constraint where

Constructors

Range :: ConstraintReason -> RealInterval Expr Expr -> Constraint

EnumeratedReal :: ConstraintReason -> [Double] -> Constraint

EnumeratedStr :: ConstraintReason -> [String] -> Constraint

data ConstraintReason

Constructors

Physical

Software

data ConstrainedChunk

ConstrainedChunks are symbolic quantities (QuantityDict) with Constraints and maybe a typical value (Maybe Expr).

Constructors

ConstrainedChunk

Fields

_qd :: QuantityDict

_constr :: [Constraint]

_reasV :: Maybe Expr

data ConstrConcept

ConstrConcepts are conceptual symbolic quantities (DefinedQuantityDict) with Constraints and maybe a reasonable value (no units!).

Constructors

ConstrConcept

Fields

_defq :: DefinedQuantityDict

_constr' :: [Constraint]

_reasV' :: Maybe Expr

cuc :: IsUnit u => String -> NP -> Symbol -> u -> Space -> [Constraint] -> Expr -> ConstrainedChunk

Creates a constrained unitary chunk from a UID, term (NP), Symbol, unit, Space, Constraints, and an Expr.

cvc :: String -> NP -> Symbol -> Space -> [Constraint] -> Maybe Expr -> ConstrainedChunk

Creates a constrained unitary chunk from a UID, term (NP), Symbol, Space, Constraints, and a Maybe Expr (Similar to cuc but no units).

constrained' :: (Concept c, MayHaveUnit c, Quantity c) => c -> [Constraint] -> Expr -> ConstrConcept

Creates a ConstrConcept with a quantitative concept, a list of Constraints and an Expr.

cuc' :: IsUnit u => String -> NP -> String -> Symbol -> u -> Space -> [Constraint] -> Expr -> ConstrConcept

Creates a constrained unitary chunk from a UID, term (NP), description (String), Symbol, unit, Space, Constraints, and an Expr.

cuc'' :: IsUnit u => String -> NP -> String -> (Stage -> Symbol) -> u -> Space -> [Constraint] -> Expr -> ConstrConcept

Similar to cuc', but Symbol is dependent on Stage.

constrainedNRV' :: (Concept c, MayHaveUnit c, Quantity c) => c -> [Constraint] -> ConstrConcept

Similar to constrained', but defaults Maybe Expr to Nothing.

cnstrw :: (Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => c -> ConstrainedChunk

Creates a new ConstrainedChunk from either a ConstrainedChunk, ConstrConcept, UncertainChunk, or an UncertQ.

cnstrw' :: (Quantity c, Concept c, Constrained c, HasReasVal c, MayHaveUnit c) => c -> ConstrConcept

Similar to cnstrw, but types must also have a Concept.

data QDefinition

A QDefinition is a QuantityDict with a defining expression (Expr), a definition (Sentence), and a domain ([UID]).

fromEqn :: IsUnit u => String -> NP -> Sentence -> Symbol -> Space -> u -> Expr -> QDefinition

Create a QDefinition with a UID, term (NP), definition (Sentence), Symbol, Space, unit, and defining expression.

fromEqn' :: String -> NP -> Sentence -> Symbol -> Space -> Expr -> QDefinition

Same as fromEqn, but has no units.

fromEqnSt :: IsUnit u => String -> NP -> Sentence -> (Stage -> Symbol) -> Space -> u -> Expr -> QDefinition

Same as fromEqn, but symbol depends on stage.

fromEqnSt' :: String -> NP -> Sentence -> (Stage -> Symbol) -> Space -> Expr -> QDefinition

Same as fromEqn, but symbol depends on stage and has no units.

equat :: Lens' QDefinition Expr

mkQuantDef :: (Quantity c, MayHaveUnit c) => c -> Expr -> QDefinition

Used to help make Qdefinitions when UID, term, and Symbol come from the same source.

mkQuantDef' :: (Quantity c, MayHaveUnit c) => c -> NP -> Expr -> QDefinition

Used to help make Qdefinitions when UID and Symbol come from the same source, with the term separate.

ec :: (Quantity c, MayHaveUnit c) => c -> Expr -> QDefinition

Smart constructor for QDefinitions. Requires a quantity and its defining equation.

data QuantityDict

QuantityDict is a combination of an IdeaDict with a quantity. Contains an IdeaDict, Space, a function from Stage -> Symbol, and Maybe a UnitDefn.

qw :: (Quantity q, MayHaveUnit q) => q -> QuantityDict

Smart constructor for a QuantityDict from another Quantity with units.

mkQuant :: String -> NP -> Symbol -> Space -> Maybe UnitDefn -> Maybe String -> QuantityDict

Make a QuantityDict from a UID, NP, Symbol, Space, Maybe UnitDefn, and an abbreviation (Maybe String).

mkQuant' :: String -> NP -> Maybe String -> Space -> (Stage -> Symbol) -> Maybe UnitDefn -> QuantityDict

Similar to mkQuant, but the abbreviation is moved to the third argument (Maybe String), and the Symbol is now dependent on Stage.

codeVC :: Idea c => c -> Symbol -> Space -> QuantityDict

Makes a QuantityDict from an Idea, Symbol, and Space. Symbol is implementation-only.

implVar :: String -> NP -> Space -> Symbol -> QuantityDict

Makes a variable that is implementation-only.

implVar' :: String -> NP -> Maybe String -> Space -> Symbol -> Maybe UnitDefn -> QuantityDict

Similar to implVar but allows specification of abbreviation and unit.

dcc :: String -> NP -> String -> ConceptChunk

Smart constructor for creating concept chunks given a UID, NounPhrase (NP) and definition (as a String).

Concept domain tagging is not yet implemented in this constructor.

dcc' :: String -> NP -> String -> String -> CommonConcept

Identical to dcc, but takes an abbreviation (String).

dccWDS :: String -> NP -> Sentence -> ConceptChunk

Similar to dcc, except the definition takes a Sentence.

dccWDS' :: String -> NP -> Sentence -> String -> CommonConcept

Similar to dcc, except the definition is a Sentence and takes an abbreviation (String).

vc :: String -> NP -> Symbol -> Space -> QuantityDict

Creates a QuantityDict from a UID, term (NP), Symbol, and Space.

vc'' :: Idea c => c -> Symbol -> Space -> QuantityDict

Creates a QuantityDict from an Idea, Symbol, and Space.

vcSt :: String -> NP -> (Stage -> Symbol) -> Space -> QuantityDict

Similar to vc, but creates a QuantityDict from something that knows about Stages.

vcUnit :: String -> NP -> Symbol -> Space -> UnitDefn -> QuantityDict

Creates a QuantityDict from a UID, term (NP), Symbol, Space, and unit (UnitDefn).

ccs :: (Idea c, Concept d) => c -> Sentence -> [d] -> ConceptChunk

Similar to cc', but allows explicit tagging.

cc :: Idea c => c -> String -> ConceptChunk

Constructor for ConceptChunk. Takes the definition of the ConceptChunk as a String. Does not allow concept domain tagging.

cc' :: Idea c => c -> Sentence -> ConceptChunk

Same as cc, except definition is a Sentence.

cic :: Concept c => String -> Sentence -> String -> c -> ConceptInstance

Constructor for a ConceptInstance. Takes in the Reference Address (String), a definition (Sentence), a short name (String), and a domain (or explicit tagging).

data UncertainChunk

UncertainChunk is a symbolic quantity with constraints, a typical value, and an uncertainty. Contains a ConstrainedChunk and an Uncertainty.

Constructors

UCh

Fields

_conc :: ConstrainedChunk

_unc' :: Uncertainty

data UncertQ

UncertQs are conceptual symbolic quantities with constraints and an Uncertainty. Contains a ConstrConcept and an Uncertainty.

uq :: (Quantity c, Constrained c, Concept c, HasReasVal c, MayHaveUnit c) => c -> Uncertainty -> UncertQ

Smart constructor that requires a Quantity, a percentage, and a typical value with an Uncertainty.

uqc :: IsUnit u => String -> NP -> String -> Symbol -> u -> Space -> [Constraint] -> Expr -> Uncertainty -> UncertQ

Uncertainty quantity (uq) but with a constraint.

uqcND :: IsUnit u => String -> NP -> Symbol -> u -> Space -> [Constraint] -> Expr -> Uncertainty -> UncertQ

Uncertainty quantity constraint (uqc) without a description.

uncrtnChunk :: (Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => c -> Uncertainty -> UncertainChunk

Smart constructor that can project to an UncertainChunk (also given an Uncertainty).

uvc :: String -> NP -> Symbol -> Space -> [Constraint] -> Expr -> Uncertainty -> UncertainChunk

Creates an uncertain variable chunk. Takes UID, term (NP), Symbol, Space, Constrains, Expr, and Uncertainty.

uncrtnw :: (HasUncertainty c, Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => c -> UncertainChunk

Projection function into an UncertainChunk from UncertQ or an UncertainChunk.

data UnitalChunk

UnitalChunks are concepts with quantities and a unit definition. Contains DefinedQuantityDicts and a UnitDefn.

Constructors

UC

Fields

_defq' :: DefinedQuantityDict

_uni :: UnitDefn

makeUCWDS :: IsUnit u => String -> NP -> Sentence -> Symbol -> u -> UnitalChunk

Creates a UnitalChunk in the same way as uc', but with a Sentence for the definition instead of a String.

uc :: (Concept c, IsUnit u) => c -> Symbol -> u -> UnitalChunk

Used to create a UnitalChunk from a Concept, Symbol, and Unit. Assumes the Space is Real.

uc' :: IsUnit u => String -> NP -> String -> Symbol -> u -> UnitalChunk

Similar to uc, except it builds the Concept portion of the UnitalChunk from a given UID, term, and definition (which are the first three arguments).

ucStaged :: IsUnit u => String -> NP -> String -> (Stage -> Symbol) -> u -> UnitalChunk

Similar to uc', but Symbol is dependent on the Stage.

ucs :: IsUnit u => String -> NP -> String -> Symbol -> Space -> u -> UnitalChunk

Similar to uc', but does not assume the Space.

ucs' :: (Concept c, IsUnit u) => c -> Symbol -> Space -> u -> UnitalChunk

Similar to uc but does not assume the Space.

ucsWS :: IsUnit u => String -> NP -> Sentence -> Symbol -> Space -> u -> UnitalChunk

Similar to ucs, but uses a Sentence for description.

class Quantity c => Unitary c where

A Unitary is a Quantity that must have a unit.

Methods

unit :: c -> UnitDefn

data UnitaryChunk

UnitaryChunks are Unitarys with Symbols. Contains a QuantityDict and a UnitDefn.

unitary :: IsUnit u => String -> NP -> Symbol -> u -> Space -> UnitaryChunk

Builds the QuantityDict part from the UID, term (NP), Symbol, and Space. Assumes there's no abbreviation.

unitary' :: IsUnit u => String -> NP -> (Stage -> Symbol) -> u -> Space -> UnitaryChunk

Same as unitary but with a Symbol that changes based on the Stage.

mkUnitary :: (Unitary u, MayHaveUnit u) => u -> UnitaryChunk

Makes a UnitaryChunk from a quantity with a unit.

unit_symb :: Unitary c => c -> USymb

Helper for getting the unit's Symbol from a chunk, as opposed to the symbols of the chunk itself.

data RelationConcept

For a ConceptChunk that also has a Relation attached.

makeRC :: UID -> NP -> Sentence -> Relation -> RelationConcept

Create a RelationConcept from a given UID, term (NP), definition (Sentence), and Relation.

addRelToCC :: Concept c => c -> UID -> Relation -> RelationConcept

Create a new RelationConcept from an old Concept. Takes a Concept, new UID and relation.

data DefinedQuantityDict

DefinedQuantityDict is the combination of a Concept and a Quantity. Contains a ConceptChunk, a Symbol dependent on Stage, a Space, and maybe a UnitDefn.

dqd :: IsUnit u => ConceptChunk -> Symbol -> Space -> u -> DefinedQuantityDict

Smart constructor that creates a DefinedQuantityDict with a ConceptChunk, a Symbol independent of Stage, a Space, and a unit.

dqd' :: ConceptChunk -> (Stage -> Symbol) -> Space -> Maybe UnitDefn -> DefinedQuantityDict

Similar to dqd, but the Symbol is now dependent on the Stage.

dqdNoUnit :: ConceptChunk -> Symbol -> Space -> DefinedQuantityDict

Similar to dqd, but without any units.

dqdQd :: (Quantity c, MayHaveUnit c) => c -> ConceptChunk -> DefinedQuantityDict

When we want to merge a quantity and a concept. This is suspicious.

dqdWr :: (Quantity c, Concept c, MayHaveUnit c) => c -> DefinedQuantityDict

When the input already has all the necessary information. A projection operator from some a type with instances of listed classes to a DefinedQuantityDict.

ucw :: (Unitary c, Concept c, MayHaveUnit c) => c -> UnitaryConceptDict

Constructs a UnitaryConceptDict from a Concept with Units.

data UnitaryConceptDict

Contains a UnitaryChunk, a definition, and a list of related UIDs.

data Derivation

Derivations are an ordered list of sentences and expressions. They are rendered in order as paragraphs and equation blocks to display the derivation.

Constructors

Derivation Sentence [Sentence]

mkDeriv :: Sentence -> [Sentence] -> Derivation

Smart constructor for Derivation

mkDerivName :: Sentence -> [Sentence] -> Derivation

Similar to mkDeriv, but prepends "Detailed derivation of" to the header

mkDerivNoHeader :: [Sentence] -> Derivation

Similar to mkDeriv, but without a header Sentence

data ShortName

shortname' :: String -> ShortName

getStringSN :: ShortName -> String

data Citation

All citations require a unique identifier used by the Drasil chunk. We will re-use the UID part as an EntryID (String) used for creating reference links. Finally we will have the reference information (CitationKind, CiteFields, and a ShortName).

type EntryID

Arguments

= String
A String that should contain no spaces.

type BibRef

Arguments

= [Citation]
A list of Citations.

citeID :: Lens' Citation UID

citeKind :: Lens' Citation CitationKind

cArticle :: People -> String -> String -> Int -> [CiteField] -> String -> Citation

Article citation requires author(s), title, journal, year. Optional fields can be: volume, number, pages, month, and note. Implicitly uses the EntryID as the chunk id.

cBookA :: People -> String -> String -> Int -> [CiteField] -> String -> Citation

Book citation by author.

Book citation requires author or editor, title, publisher, year. Optional fields can be volume or number, series, address, edition, month, and note. Implicitly uses the EntryID as the chunk id.

cBookE :: People -> String -> String -> Int -> [CiteField] -> String -> Citation

Book citation by editor.

Book citation requires author or editor, title, publisher, year. Optional fields can be volume or number, series, address, edition, month, and note. Implicitly uses the EntryID as the chunk id.

cBooklet :: String -> [CiteField] -> String -> Citation

Booklet citation requires title. Optional fields can be author, how published, address, month, year, note. Implicitly uses the EntryID as the chunk id.

cInBookACP :: People -> String -> Int -> [Int] -> String -> Int -> [CiteField] -> String -> Citation

InBook citation by author.

InBook citation requires author or editor, title, chapter and/or pages, publisher, year. Optional fields can be volume or number, series, type, address, edition, month, and note. Implicitly uses the EntryID as the chunk id. This smart constructor includes both chapter and page numbers.

cInBookECP :: People -> String -> Int -> [Int] -> String -> Int -> [CiteField] -> String -> Citation

InBook citation by editor.

InBook citation requires author or editor, title, chapter and/or pages, publisher, year. Optional fields can be volume or number, series, type, address, edition, month, and note. Implicitly uses the EntryID as the chunk id. This smart constructor includes both chapter and page numbers.

cInBookAC :: People -> String -> Int -> String -> Int -> [CiteField] -> String -> Citation

Otherwise identical to cInBookACP.

InBook citation excluding page numbers.

cInBookEC :: People -> String -> Int -> String -> Int -> [CiteField] -> String -> Citation

Otherwise identical to cInBookECP.

InBook citation excluding page numbers.

cInBookAP :: People -> String -> [Int] -> String -> Int -> [CiteField] -> String -> Citation

Otherwise identical to cInBookACP.

InBook citation excluding chapter.

cInBookEP :: People -> String -> [Int] -> String -> Int -> [CiteField] -> String -> Citation

Otherwise identical to cInBookECP.

InBook citation excluding chapter.

cInCollection :: People -> String -> String -> String -> Int -> [CiteField] -> String -> Citation

InCollection citation requires author, title, bookTitle, publisher, year. Optional fields can be editor, volume or number, series, type, chapter, pages, address, edition, month, and note. Implicitly uses the EntryID as the chunk id.

cInProceedings :: People -> String -> String -> Int -> [CiteField] -> String -> Citation

InProceedings citation requires author, title, bookTitle, year. Optional fields can be editor, volume or number, series, pages, address, month, organization, publisher, and note. Implicitly uses the EntryID as the chunk id.

cManual :: String -> [CiteField] -> String -> Citation

Manual (technical documentation) citation requires title. Optional fields can be author, organization, address, edition, month, year, and note. Implicitly uses the EntryID as the chunk id.

cMThesis :: People -> String -> String -> Int -> [CiteField] -> String -> Citation

Master's Thesis citation requires author, title, school, and year. Optional fields can be type, address, month, and note. Implicitly uses the EntryID as the chunk id.

cMisc :: [CiteField] -> String -> Citation

Misc citation requires nothing. Optional fields can be author, title, howpublished, month, year, and note. Implicitly uses the EntryID as the chunk id.

cPhDThesis :: People -> String -> String -> Int -> [CiteField] -> String -> Citation

PhD Thesis citation requires author, title, school, and year. Optional fields can be type, address, month, and note. Implicitly uses the EntryID as the chunk id.

cProceedings :: String -> Int -> [CiteField] -> String -> Citation

Proceedings citation requires title and year. Optional fields can be editor, volume or number, series, address, publisher, note, month, and organization. Implicitly uses the EntryID as the chunk id.

cTechReport :: People -> String -> String -> Int -> [CiteField] -> String -> Citation

Technical Report citation requires author, title, institution, and year. Optional fields can be type, number, address, month, and note. Implicitly uses the EntryID as the chunk id.

cUnpublished :: People -> String -> String -> [CiteField] -> String -> Citation

Unpublished citation requires author, title, and note. Optional fields can be month and year. Implicitly uses the EntryID as the chunk id.

class HasCitation c where

Some documents, as well as some pieces of knowledge, have citations

Methods

getCitations :: Lens' c [Citation]

data Sentence where

Constructors

Ch :: SentenceStyle -> UID -> Sentence

Sy :: USymb -> Sentence

S

Fields

:: String

-> Sentence

Strings, used for Descriptions in Chunks

P :: Symbol -> Sentence

E :: Expr -> Sentence

Ref :: Reference -> Sentence

Quote

Fields

:: Sentence

-> Sentence

Adds quotation marks around a sentence

Percent

Fields

:: Sentence

% symbol

(:+:) :: Sentence -> Sentence -> Sentence infixr 5
Direct concatenation of two Sentences (no implicit spaces!)

EmptyS :: Sentence

data SentenceStyle

For writing "sentences" via combining smaller elements. Sentences are made up of some known vocabulary of things:

units (their visual representation)

words (via String)

special characters

accented letters

References to specific layout objects

Constructors

PluralTerm

SymbolStyle

TermStyle

ShortStyle

(+:+) :: Sentence -> Sentence -> Sentence

Helper for concatenating two sentences with a space between them.

(+:+.) :: Sentence -> Sentence -> Sentence

Helper which concatenates two sentences using +:+ then adds a period to the end.

(+:) :: Sentence -> Sentence -> Sentence

Helper which concatenates two sentences using +:+ then adds a colon to the end.

(!.) :: Sentence -> Sentence

Helper which appends a period to the end of a sentence

capSent :: Sentence -> Sentence

Capitalizes a Sentence.

ch :: (HasUID c, HasSymbol c) => c -> Sentence

Gets a symbol and places in a Sentence

sC :: Sentence -> Sentence -> Sentence

Helper for concatenating two sentences with a comma and space between them.

sDash :: Sentence -> Sentence -> Sentence

Helper for concatenating two sentences with a space-surrounded dash between them.

sParen :: Sentence -> Sentence

Helper for wrapping sentences in parentheses.

sdep :: Sentence -> [UID]

shortdep :: Sentence -> [UID]

data Reference

Constructors

Reference UID LblType ShortName RefInfo

data RefInfo

Constructors

None

Equation [Int]

Page [Int]

RefNote String

class NounPhrase n where

Methods

phraseNP

Arguments

:: n

-> Sentence
ex. "the quick brown fox"

pluralNP

Arguments

:: n

-> PluralForm
ex. "the quick brown foxes" Could replace plural string with a function.

sentenceCase

Arguments

:: n

-> (NP -> Sentence)

-> Capitalization
example: "The quick brown fox"

titleCase

Arguments

:: n

-> (NP -> Sentence)

-> Capitalization
ex. "The Quick Brown Fox"

data NP

For nouns and NounPhrases

pn :: String -> NP

Self plural.

Construct a Proper Noun, it is always capitalized as written.

pn' :: String -> NP

Plural form simply adds "s" (ex. Henderson -> Hendersons).

Construct a Proper Noun, it is always capitalized as written.

pn'' :: String -> NP

Plural form adds "e".

Construct a Proper Noun, it is always capitalized as written.

pn''' :: String -> NP

Plural form adds "es" (ex. Bush -> Bushes).

Construct a Proper Noun, it is always capitalized as written.

pnIrr :: String -> PluralRule -> NP

Construct a Proper Noun. Meant for use with IrregPlur from PluralRule. First argument is the String representing the noun, second is the rule.

cn :: String -> NP

Self plural.

Construct a common noun which capitalizes the first letter of the first word at the beginning of a sentence.

cn' :: String -> NP

Plural form simply adds "s" (ex. dog -> dogs).

Construct a common noun which capitalizes the first letter of the first word at the beginning of a sentence.

cn'' :: String -> NP

Plural form adds "e" (ex. formula -> formulae).

Construct a common noun which capitalizes the first letter of the first word at the beginning of a sentence.

cn''' :: String -> NP

Plural form adds "es" (ex. bush -> bushes).

Construct a common noun which capitalizes the first letter of the first word at the beginning of a sentence.

cnIP :: String -> PluralRule -> NP

Common noun that allows you to specify the pluralization rule (as in pnIrr).

cnIrr :: String -> PluralRule -> CapitalizationRule -> NP

Common noun that allows you to specify both the pluralization rule and the capitalization rule for sentence case (if the noun is used at the beginning of a sentence).

cnIES :: String -> NP

Common noun that pluralizes by dropping the last letter and adding an "ies" ending (ex. body -> bodies).

cnICES :: String -> NP

Common noun that pluralizes by dropping the last two letters and adding an "ices" ending (ex. matrix -> matrices).

cnIS :: String -> NP

Common noun that pluralizes by dropping the last two letters and adding "es" (ex. analysis -> analyses).

cnUM :: String -> NP

Common noun that pluralizes by dropping the last two letters and adding "a" (ex. datum -> data).

nounPhrase :: String -> PluralString -> NP

Creates a NP with a given singular and plural form that capitalizes the first letter of the first word for sentence case.

nounPhrase' :: String -> PluralString -> CapitalizationRule -> NP

Similar to nounPhrase, but following a given capitalization rule for sentence case.

data CapitalizationRule

Capitalization rules.

Constructors

CapFirst
Capitalize the first letter of the first word only.

CapWords
Capitalize the first letter of each word.

Replace Sentence
Replace the noun phrase with the given Sentence. Used for custom capitalization.

atStartNP :: NounPhrase n => n -> Capitalization

Singular sentence case.

Helper function for getting the sentence case of a noun phrase.

atStartNP' :: NounPhrase n => n -> Capitalization

Plural sentence case.

Helper function for getting the sentence case of a noun phrase.

data PluralRule

Pluralization rules.

Constructors

AddS
Add "s" to the end of the noun phrase.

AddE
Add "e" to the end of the noun phrase.

AddES
Add "es" to the end of the noun phrase.

SelfPlur
The noun phrase is already plural.

IrregPlur (String -> String)
Apply the given function to the noun phrase to get the plural.

compoundPhrase :: (NounPhrase a, NounPhrase b) => a -> b -> NP

Combine two noun phrases. The singular form becomes phrase from t1 followed by phrase of t2. The plural becomes phrase of t1 followed by plural of t2. Uses standard CapFirst sentence case and CapWords title case. For example: compoundPhrase system constraint will have singular form "system constraint" and plural "system constraints".

compoundPhrase' :: NP -> NP -> NP

Similar to compoundPhrase, but where the sentence case is the same as the title case (CapWords).

compoundPhrase'' :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Similar to compoundPhrase', but which accepts functions to be used for constructing the plural form. For example compoundPhrase'' plural phrase system constraint would have the plural form "systems constraint".

compoundPhrase''' :: (NP -> Sentence) -> NP -> NP -> NP

Used when you need a special function apllied to the first term of both cases (eg. short or plural).

compoundPhraseP1 :: NP -> NP -> NP

titleizeNP :: NounPhrase n => n -> Capitalization

Singular title case.

Helper function for getting the title case of a noun phrase.

titleizeNP' :: NounPhrase n => n -> Capitalization

Plural title case.

Helper function for getting the title case of a noun phrase.

nounPhrase'' :: Sentence -> PluralForm -> CapitalizationRule -> CapitalizationRule -> NP

Custom noun phrase constructor that takes a singular form, plural form, sentence case capitalization rule, and title case capitalization rule.

nounPhraseSP :: String -> NP

For things that should not be pluralized. Works like nounPhrase, but with only the first argument.

nounPhraseSent :: Sentence -> NP

For Reuirements, Assumptions, LikelyChanges, etc. to allow for referencing.

class HasUID s => Referable s where

Has the ability to be referenced

Methods

refAdd

Arguments

:: s

-> String
The referencing address (what we're linking to). Only visible in the source (tex/html).

renderRef

Arguments

:: s

-> LblType
alternate form

data Document

A Document has a Title (Sentence), Author(s) (Sentence), and Sections which hold the contents of the document

Constructors

Document Title Author [Section]

data DType

Types of definitions

Constructors

General

Instance

Theory

Data

data Section

Sections have a title (Sentence) and a list of contents (SecCons) and its shortname

Constructors

Section

Fields

tle :: Title

cons :: [SecCons]

_lab :: Reference

data Contents

Constructors

UlC UnlabelledContent

LlC LabelledContent

data SecCons

Section Contents are split into subsections or contents, where contents are standard layout objects (see Contents)

Constructors

Sub Section

Con Contents

data ListType

Constructors

Bullet [(ItemType, Maybe String)]
Bulleted list

Numeric [(ItemType, Maybe String)]
Enumerated List

Simple [ListTuple]
Simple list with items denoted by -

Desc [ListTuple]
Descriptive list, renders as "Title: Item" (see ListTuple)

Definitions [ListTuple]
Renders a list of "Title is the Item"

data ItemType

Constructors

Flat Sentence
Standard singular item

Nested Header ListType
Nest a list as an item

type ListTuple

Arguments

= (Title, ItemType, Maybe String)
Title: Item

data LabelledContent

Contains a Reference and RawContent

Constructors

LblC

Fields

_ref :: Reference

_ctype :: RawContent

newtype UnlabelledContent

Only contains RawContent

Constructors

UnlblC

Fields

_cntnts :: RawContent

extractSection :: Document -> [Section]

mkParagraph :: Sentence -> Contents

Smart constructor needed for LabelledContent

mkRawLC :: RawContent -> Reference -> LabelledContent

Smart constructor needed for LabelledContent

llcc :: Reference -> RawContent -> LabelledContent

Smart constructor for labelled content chunks

ulcc :: RawContent -> UnlabelledContent

Smart constructor for unlabelled content chunks

section :: Sentence -> [Contents] -> [Section] -> Reference -> Section

Smart constructor for creating Sections with introductory contents (ie. paragraphs, tables, etc.) and a list of subsections.

fig :: Lbl -> Filepath -> RawContent

Figure smart constructor. Assumes 100% of page width as max width.

figWithWidth :: Lbl -> Filepath -> MaxWidthPercent -> RawContent

Figure smart constructor for customized max widths.

type MaxWidthPercent = Float

MaxWidthPercent should be kept in the range 1-100. Values outside this range may have unexpected results. Used for specifying max figure width as pagewidth*MaxWidthPercent/100.

class HasContents c where

Methods

accessContents :: Lens' c RawContent

data RawContent

Types of layout objects we deal with explicitly

Constructors

Table [Sentence] [[Sentence]] Title Bool
table has: header-row data(rows) label/caption showlabel?

Paragraph Sentence
Paragraphs are just sentences.

EqnBlock Expr

DerivBlock Sentence [RawContent]

Enumeration ListType
Lists

Defini DType [(Identifier, [Contents])]

Figure Lbl Filepath MaxWidthPercent
Should use relative file path.

Bib BibRef

Graph [(Sentence, Sentence)] (Maybe Width) (Maybe Height) Lbl
TODO: Fill this one in.

mkFig :: Reference -> RawContent -> Contents

Smart constructor needed for LabelledContent

makeTabRef :: String -> Reference

makeFigRef :: String -> Reference

makeSecRef :: String -> String -> Reference

makeLstRef :: String -> String -> Reference

makeURI :: UID -> String -> ShortName -> Reference

Create a reference for a URI

data Space

Spaces

Constructors

Integer

Rational

Real

Natural

Boolean

Char

String

Radians

Vect Space

Array Space

Actor String

DiscreteD [Double]

DiscreteS [String]

Void

data RealInterval a b where

RealInterval. A |RealInterval| is a subset of |Real| (as a |Space|). These come in different flavours. For now, embed |Expr| for the bounds, but that will change as well.

Constructors

Bounded :: (Inclusive, a) -> (Inclusive, b) -> RealInterval a b

UpTo :: (Inclusive, a) -> RealInterval a b

UpFrom :: (Inclusive, b) -> RealInterval a b

data Inclusive

Constructors

Inc

Exc

data RTopology

Topology of a subset of reals.

Constructors

Continuous

Discrete

data DomainDesc a b where

Constructors

BoundedDD :: Symbol -> RTopology -> a -> b -> DomainDesc a b

AllDD :: Symbol -> RTopology -> DomainDesc a b

getActorName :: Space -> String

getInnerSpace :: Space -> Space

data Decoration

Decorations on symbols/characters such as hats or Vector representations (bolding/etc)

Constructors

Hat

Vector

Prime

data Symbol

A Symbol is actually going to be a graphical description of what gets rendered as a (unique) symbol. This is actually NOT based on semantics at all, but just a description of how things look.

Symbols can be:

Variable (string such as "x" that represent a value that can vary)

Label (strings such as "max" or "target" that represent a single idea)

Special characters (ex. unicode)

Decorated symbols

Concatenations of symbols, including subscripts and superscripts

Empty! (this is to give this a monoid-like flavour)

Constructors

Variable String

Label String

Integ Int

Special Special

Atop Decoration Symbol

Corners [Symbol] [Symbol] [Symbol] [Symbol] Symbol
In order: upleft lowleft upright lowright base

Corners [1]   [2]   [3]   [4]   [5]

 Visually:  [1]   [3]

               [5]

            [2]   [4]

Concat [Symbol]
[s1, s2] -> s1s2

Empty

compsy :: Symbol -> Symbol -> Ordering

The default compare function sorts all the lower case after the upper case. Comparation is used twice for each Atomic case, once for making sure they are the same letter, once for case sensitive. As far as this comparison is considered, Δ is a "decoration" and ignored unless the compared symbols are the exact same, in which case it is ordered after the undecorated symbol.

mkTable :: [a -> b] -> [a] -> [[b]]

Create a table body (not including header row) by applying the given functions to the column elements of the table rows (in order). The first argument is a list of functions to be applied (one per column). This essentially creates the rows. The second argument is a list of elements apply the functions to.

For example, mkTable [id, *5] [1,2,3] should produce a table:

| 1 |  5 |
| 2 | 10 |
| 3 | 15 |

type People = [Person]

data Person

A person can have a given name, middle name(s), and surname, as well as the naming convention they use.

person :: String -> String -> Person

Constructor for a person using Western naming conventions. Used for a person with only a given name and surname. Arguments are in the order: given name, surname.

class HasName p

Minimal complete definition

nameStr

name :: HasName n => n -> String

person' :: String -> String -> Person

Constructor for a person using Eastern naming conventions. Used for a person with only a given name and surname. Arguments are in the order: surname, given name

personWM :: String -> [String] -> String -> Person

Constructor for a person using Western naming conventions. Similar to the person constructor, except the middle argument is a list of middle names.

personWM' :: String -> [String] -> String -> Person

Constructor for a person using Eastern naming conventions. Similar to the 'person'' constructor, except the middle argument is a list of middle names.

mononym :: String -> Person

Constructor for a person with a mononym (only one name).

nameStr :: HasName p => p -> String

rendPersLFM :: Person -> String

rendPersLFM' :: Person -> String

rendPersLFM'' :: Person -> String

comparePeople :: [Person] -> [Person] -> Ordering

data Stage

Stages correspond to what we're trying to look up. They range from abstract to concrete.

Constructors

Equational

Implementation

eqSymb :: HasSymbol q => q -> Symbol

Helper function for getting a symbol in the Equational Stage

codeSymb :: HasSymbol q => q -> Symbol

Helper function for getting a symbol in the Implementation Stage

hasStageSymbol :: HasSymbol q => q -> Stage -> Bool

Is a Stage symbol real or Empty?

autoStage :: Symbol -> Stage -> Symbol

Helper for creating a symbol with Unicode in it.

hat :: Symbol -> Symbol

Helper for creating a symbol with a hat ("^") atop it.

prime :: Symbol -> Symbol

Helper for creating a Vector symbol.

staged :: Symbol -> Symbol -> Stage -> Symbol

Helper for creating a symbol that depends on the stage

sub :: Symbol -> Symbol -> Symbol

Helper for creating a symbol with a subscript to the right. Arguments: Base symbol, then subscripted symbol.

subStr :: Symbol -> String -> Symbol

Helper for a common case of subscript, with a string Arguments: Base symbol, then subscript String.

sup :: Symbol -> Symbol -> Symbol

Helper for creating a symbol with a superscript to the right. Arguments: Base symbol, then superscripted symbol.

unicodeConv :: Symbol -> Symbol

Helper for autoStage that apples unicodeString to all Symbols with Strings

upperLeft :: Symbol -> Symbol -> Symbol

Helper for creating a symbol with a superscript on the left side of the symbol. Arguments: Base symbol, then superscripted symbol.

vec :: Symbol -> Symbol

Helper for creating a Vector symbol.

makeRef2S :: (Referable l, HasShortName l) => l -> Sentence

makeCite :: Citation -> Reference

makeCiteS :: Citation -> Sentence

makeRef2 :: (Referable l, HasShortName l) => l -> Reference

makeCiteInfo :: Citation -> RefInfo -> Reference

makeCiteInfoS :: Citation -> RefInfo -> Sentence

getAdd :: LblType -> String

Retrieves the String contained in a LblType

prepend :: String -> IRefProg

Prepends a String to an IRefProg

data LblType

Applying different pieces of information for a reference. An RP is a decorated internal reference. Citation is a citation. URI is for URLs and other external links.

Constructors

RP IRefProg String

Citation String

URI String

data IRefProg

Used in LblType

Constructors

Deferred UID
Deferred lookup; done later

RS String
Lifts a String into a RefProg

RConcat IRefProg IRefProg
Concatenates with two subprograms

Name
The Symbol to insert the ShortName directly

introduceAbb :: Idea n => n -> Sentence

Helper for common pattern of introducing the title-case version of a noun phrase (from an Idea) followed by its abbreviation in parentheses.

phrase :: (HasUID n, NamedIdea n) => n -> Sentence

Helper for getting the phrase from a NamedIdea.

plural :: (HasUID n, NamedIdea n) => n -> Sentence

Helper for getting the plural of a phrase from a NamedIdea

phrasePoss :: NamedIdea n => n -> Sentence

Singular possesive function

pluralPoss :: NamedIdea n => n -> Sentence

Plural possesive function

atStart :: NamedIdea n => n -> Sentence

Singular sentence case.

Helper function for getting the sentence case of a noun phrase from a NamedIdea.

atStart' :: NamedIdea n => n -> Sentence

Plural sentence case.

Helper function for getting the sentence case of a noun phrase from a NamedIdea.

titleize :: NamedIdea n => n -> Sentence

Singular title case.

Helper function for getting the title case of a noun phrase from a NamedIdea.

titleize' :: NamedIdea n => n -> Sentence

Plural title case.

Helper function for getting the title case of a noun phrase from a NamedIdea.

data Uncertainty

May contain an uncertainty value and Maybe Int

uncty :: Double -> Maybe Int -> Uncertainty

Smart constructor for values with uncertainty

defaultUncrt :: Uncertainty

uncVal :: HasUncertainty x => x -> Double

uncPrec :: HasUncertainty x => x -> Maybe Int

exact :: Uncertainty

Smart constructor for exact values

newtype USymb

Language of units (how to build them up) UName for the base cases, otherwise build up. Probably a 7-vector would be better (less error-prone!) | Language of unit equations, to define a unit relative to another

Constructors

US [(Symbol, Integer)]

data Month

Month abbreviations

Constructors

Jan

Feb

Mar

Apr

May

Jun

Jul

Aug

Sep

Oct

Nov

Dec

data CiteField

Fields used in citations.

Constructors

Address String

Author People

BookTitle String
Used for InCollection references only.

Chapter Int

Edition Int

Editor People

HowPublished HP

Institution String

Journal String

Month Month

Note String

Number Int

Organization String

Pages [Int]
Range of pages (ex1. 1-32; ex2. 7,31,52-55)

Publisher String

School String

Series String

Title String

Type String
BibTeX "type" field

Volume Int

Year Int

data HP

How Published. Necessary for URLs to work properly.

Constructors

URL String

Verb String

data CitationKind

External references come in many flavours. Articles, Books, etc. (we are using the types available in Bibtex)

Constructors

Article

Book

Booklet

InBook

InCollection

InProceedings

Manual

MThesis

Misc

PhDThesis

Proceedings

TechReport

Unpublished

author :: People -> CiteField

Smart field constructor

editor :: People -> CiteField

Smart field constructor

address :: String -> CiteField

Smart field constructor

bookTitle :: String -> CiteField

Smart field constructor

howPublished :: String -> CiteField

Smart field constructor

howPublishedU :: String -> CiteField

Smart field constructor

institution :: String -> CiteField

Smart field constructor

journal :: String -> CiteField

Smart field constructor

note :: String -> CiteField

Smart field constructor

organization :: String -> CiteField

Smart field constructor

publisher :: String -> CiteField

Smart field constructor

school :: String -> CiteField

Smart field constructor

series :: String -> CiteField

Smart field constructor

title :: String -> CiteField

Smart field constructor

typeField :: String -> CiteField

Smart field constructor

chapter :: Int -> CiteField

Smart field constructor

edition :: Int -> CiteField

Smart field constructor

number :: Int -> CiteField

Smart field constructor

volume :: Int -> CiteField

Smart field constructor

year :: Int -> CiteField

Smart field constructor

pages :: [Int] -> CiteField

Smart field constructor

month :: Month -> CiteField

Smart field constructor

data UnitDefn

For defining units. It has a ConceptChunk (that defines what kind of unit it is), a unit symbol, maybe another (when it is a synonym), perhaps a definition, and a list of UID of the units that make up the definition.

Constructors

UD

Fields

_vc :: ConceptChunk

_cas :: UnitSymbol

_cu :: [UID]

fromUDefn :: UDefn -> USymb

Can generate a default symbol

unitCon :: String -> ConceptChunk

Helper for fundamental unit concept chunk creation. Uses the same String for the identifier, term, and definition.

makeDerU :: ConceptChunk -> UnitEquation -> UnitDefn

Create a derived unit chunk from a concept and a unit equation.

(^:) :: UnitDefn -> Integer -> UnitEquation

Combinator for raising a unit to a power.

(/:) :: UnitDefn -> UnitDefn -> UnitEquation

Combinator for dividing one unit by another.

(*:) :: UnitDefn -> UnitDefn -> UnitEquation

Combinator for multiplying two units together.

(*$) :: UnitDefn -> UnitEquation -> UnitEquation

Combinator for multiplying a unit and a symbol.

(/$) :: UnitDefn -> UnitEquation -> UnitEquation

Combinator for dividing a unit and a symbol.

(^$) :: UnitEquation -> UnitEquation -> UnitEquation

Combinator for mulitiplying two unit equations.

newUnit :: String -> UnitEquation -> UnitDefn

Smart constructor for new derived units from existing units.

scale :: IsUnit s => Double -> s -> UDefn

Combinator for scaling one unit by some number.

shift :: IsUnit s => Double -> s -> UDefn

Combinator for shifting one unit by some number.

derUC :: String -> String -> String -> Symbol -> UDefn -> UnitDefn

Uses self-plural term.

Create a derived unit chunk from a UID, term (String), definition, Symbol, and unit equation.

derUC' :: String -> String -> String -> Symbol -> UDefn -> UnitDefn

Uses term that pluralizes by adding "s" to the end.

Create a derived unit chunk from a UID, term (String), definition, Symbol, and unit equation.

derUC'' :: String -> NP -> String -> Symbol -> UDefn -> UnitDefn

Create a derived unit chunk from a UID, term (NP), definition, Symbol, and unit equation.

fund :: String -> String -> String -> UnitDefn

Smart constructor for a "fundamental" unit.

fund' :: String -> String -> Symbol -> UnitDefn

Variant of the fund, useful for degree.

compUnitDefn :: UnitDefn -> UnitDefn -> Ordering

We don't want an Ord on units, but this still allows us to compare them.

derCUC :: String -> String -> String -> Symbol -> UnitEquation -> UnitDefn

Create a SI_Unit with two Symbol representations. The created NP is self-plural.

derCUC' :: String -> String -> String -> Symbol -> UnitEquation -> UnitDefn

Similar to derCUC, but the created NP has the AddS plural rule.

derCUC'' :: String -> NP -> String -> Symbol -> UnitEquation -> UnitDefn

Create a derived unit chunk from a UID, term (NP), definition, Symbol, and unit equation.

unitWrapper :: IsUnit u => u -> UnitDefn

For allowing lists to mix the two, thus forgetting the definition part.

getCu :: UnitEquation -> [UID]

Get a list of UID of the units that make up the UnitEquation.

class MayHaveUnit u where

Types may contain a unit (UnitDefn).

Methods

getUnit :: u -> Maybe UnitDefn