Language.Drasil.Chunk.CommonIdea

Description

Contains the common idea type and respective constructors.

Common Idea datatype

data CI

The common idea (with NounPhrase) data type. It must have a UID, NounPhrase for its term, an abbreviation (String), and a domain ([UID]). It is similar to IdeaDict and IdeaDict in the sense that these are for things worth naming, but this type also carries an abbreviation and related domains of knowledge.

Ex. The term "Operating System" has the abbreviation OS and comes from the domain of computer science.

Constructors

commonIdea :: String -> NP -> String -> [UID] -> CI

The commonIdea smart constructor requires a chunk id (String), a term (NP), an abbreviation (String), and a domain ([UID]).

commonIdeaWithDict :: String -> NP -> String -> [IdeaDict] -> CI

Similar to commonIdea, but takes a list of IdeaDict (often a domain).

Functions

prependAbrv :: CommonIdea c => c -> String -> String

Prepends the abbreviation from a CommonIdea to a String.