Language.Drasil.Chunk.Concept

Documentation

data ConceptChunk

The ConceptChunk datatype is a Concept that contains an idea (IdeaDict), a definition (Sentence), and a domain ([UID]).

dcc :: String -> NP -> String -> ConceptChunk

Smart constructor for creating concept chunks given a UID, NounPhrase (NP) and definition (as a String).

Concept domain tagging is not yet implemented in this constructor.

dcc' :: String -> NP -> String -> String -> CommonConcept

Identical to dcc, but takes an abbreviation (String).

dccWDS :: String -> NP -> Sentence -> ConceptChunk

Similar to dcc, except the definition takes a Sentence.

dccWDS' :: String -> NP -> Sentence -> String -> CommonConcept

Similar to dcc, except the definition is a Sentence and takes an abbreviation (String).

cc :: Idea c => c -> String -> ConceptChunk

Constructor for ConceptChunk. Takes the definition of the ConceptChunk as a String. Does not allow concept domain tagging.

cc' :: Idea c => c -> Sentence -> ConceptChunk

Same as cc, except definition is a Sentence.

ccs :: (Idea c, Concept d) => c -> Sentence -> [d] -> ConceptChunk

Similar to cc', but allows explicit tagging.

cic :: Concept c => String -> Sentence -> String -> c -> ConceptInstance

Constructor for a ConceptInstance. Takes in the Reference Address (String), a definition (Sentence), a short name (String), and a domain (or explicit tagging).

cw :: Concept c => c -> ConceptChunk

For projecting out to the ConceptChunk data-type.

data CommonConcept

Contains a common idea (CI) with a definition (Sentence).

data ConceptInstance

Contains a ConceptChunk, reference address, and a ShortName.