Language.Drasil.Chunk.Constrained

Description

Add constraints and a reasonable value to chunks that are quantities.

Constrained Chunks

From a Concept

data ConstrConcept

ConstrConcepts are conceptual symbolic quantities (DefinedQuantityDict) with Constraints and maybe a reasonable value (no units!). Similar to ConstrainedChunk but includes a definition and domain.

Ex. Measuring the length of a pendulum arm could be a concept that has some reasonable value (between 1 cm and 2 m) and the constraint that the length cannot be a negative value.

Constructors

ConstrConcept

Fields

_defq :: DefinedQuantityDict

_constr' :: [ConstraintE]

_reasV' :: Maybe Expr

cnstrw' :: (Quantity c, Concept c, Constrained c, HasReasVal c, MayHaveUnit c) => c -> ConstrConcept

Similar to cnstrw, but types must also have a Concept.

constrained' :: (Concept c, MayHaveUnit c, Quantity c) => c -> [ConstraintE] -> Expr -> ConstrConcept

Creates a ConstrConcept with a quantitative concept, a list of Constraints and an Expr.

constrainedNRV' :: (Concept c, MayHaveUnit c, Quantity c) => c -> [ConstraintE] -> ConstrConcept

Similar to constrained', but defaults Maybe Expr to Nothing.

cuc' :: IsUnit u => String -> NP -> String -> Symbol -> u -> Space -> [ConstraintE] -> Expr -> ConstrConcept

Creates a constrained unitary chunk from a UID, term (NP), description (String), Symbol, unit, Space, Constraints, and an Expr.

cuc'' :: IsUnit u => String -> NP -> String -> (Stage -> Symbol) -> u -> Space -> [ConstraintE] -> Expr -> ConstrConcept

Similar to cuc', but Symbol is dependent on Stage.

cucNoUnit' :: String -> NP -> String -> Symbol -> Space -> [ConstraintE] -> Expr -> ConstrConcept

Similar to cuc', but does not include a unit.