Language.Drasil.Chunk.DefinedQuantity

Description

Contains types that define quantities from concepts. Similar to QuantityDict.

Chunk Type

data DefinedQuantityDict

DefinedQuantityDict is the combination of a Concept and a Quantity. Contains a ConceptChunk, a Symbol dependent on Stage, a Space, and maybe a UnitDefn. Used when we want to assign a quantity to a concept. Includes the space, symbol, and units for that quantity.

Ex. A pendulum arm can be defined as a concept with a symbol (l), space (Real numbers), and units (cm, m, etc.).

Constructors

dqd :: IsUnit u => ConceptChunk -> Symbol -> Space -> u -> DefinedQuantityDict

Smart constructor that creates a DefinedQuantityDict with a ConceptChunk, a Symbol independent of Stage, a Space, and a unit.

dqdNoUnit :: ConceptChunk -> Symbol -> Space -> DefinedQuantityDict

Similar to dqd, but without any units.

dqd' :: ConceptChunk -> (Stage -> Symbol) -> Space -> Maybe UnitDefn -> DefinedQuantityDict

Similar to dqd, but the Symbol is now dependent on the Stage.

dqdQd :: (Quantity c, MayHaveUnit c) => c -> ConceptChunk -> DefinedQuantityDict

When we want to merge a quantity and a concept. This is suspicious.

dqdWr :: (Quantity c, Concept c, MayHaveUnit c) => c -> DefinedQuantityDict

When the input already has all the necessary information. A projection operator from some a type with instances of listed classes to a DefinedQuantityDict.

tempdqdWr' :: (Quantity c, Concept c, MayHaveUnit c) => c -> DefinedQuantityDict

Temporary projection constructor, not to be used outside drasil-lang.