Language.Drasil.Chunk.DifferentialModel

Export Data Type

data DifferentialModel

Describe the structural content of a system of linear ODEs with six necessary fields

Constructors

SystemOfLinearODEs

Fields

_indepVar :: UnitalChunk

independent variable, often time

_depVar :: ConstrConcept

dependent variable

_coefficients :: [[Expr]]

coefficients matrix

_unknowns :: [Unknown]

unknowns column vector (orders)

_dmConstants :: [Expr]

constant column vector

_dmconc :: ConceptChunk

meta data

data ODESolverFormat

Acceptable format for ODE solvers, represent the structure of X' = AX + B X' is a column vector of first-order unknowns

Constructors

X'

Fields

coeffVects :: [[Expr]]

represent A, the coefficient matrix with identity matrix

unknownVect :: [Integer]

combing with the dependent variable. it represents X, the unknown column vector after reduce the highest order.

constantVect :: [Expr]

represent B, the constant column vector with identity matrix

data InitialValueProblem

Information for solving an initial value problem

Constructors

IVP

Fields

initTime :: Expr

initial time

finalTime :: Expr

end time

initValues :: [Expr]

initial values

Input Language

($^^) :: ConstrConcept -> Integer -> Unknown

Operation connect the dependent variable and the order

($**) :: Expr -> Unknown -> Term

Operation represent multiple

($++) :: [Term] -> Term -> LHS

Operation represent plus (collection Terms)

Constructors

makeAODESolverFormat :: DifferentialModel -> ODESolverFormat

Construct an ODESolverFormat for solving the ODE.

makeAIVP :: Expr -> Expr -> [Expr] -> InitialValueProblem

makeASystemDE :: UnitalChunk -> ConstrConcept -> [[Expr]] -> [Unknown] -> [Expr] -> String -> NP -> Sentence -> DifferentialModel

Create a DifferentialModel by giving a independent variable, a dependent variable a canonical matrix form, and conceptChuck.

makeASingleDE :: UnitalChunk -> ConstrConcept -> LHS -> Expr -> String -> NP -> Sentence -> DifferentialModel

Create a DifferentialModel by the input language

formEquations :: [[Expr]] -> [Unknown] -> [Expr] -> ConstrConcept -> [Expr]

Form well-formatted ODE equations which the ODE solvers can solve.