Language.Drasil.Chunk.Quantity

Description

Defines chunks that add quantities to an idea. Similar to DefinedQuantityDict.

Chunk Type

data QuantityDict

QuantityDict is a combination of an IdeaDict with a quantity. Contains an IdeaDict, Space, a function from Stage -> Symbol, and Maybe a UnitDefn.

Ex. A pendulum arm does not necessarily have to be defined as a concept before we assign a space (Real numbers), a symbol (l), or units (cm, m, etc.).

Constructors

codeVC :: Idea c => c -> Symbol -> Space -> QuantityDict

Makes a QuantityDict from an Idea, Symbol, and Space. Symbol is implementation-only.

implVar :: String -> NP -> Space -> Symbol -> QuantityDict

Makes a variable that is implementation-only.

implVar' :: String -> NP -> Maybe String -> Space -> Symbol -> Maybe UnitDefn -> QuantityDict

Similar to implVar but allows specification of abbreviation and unit.

mkQuant :: String -> NP -> Symbol -> Space -> Maybe UnitDefn -> Maybe String -> QuantityDict

Make a QuantityDict from a UID, NP, Symbol, Space, Maybe UnitDefn, and an abbreviation (Maybe String).

mkQuant' :: String -> NP -> Maybe String -> Space -> (Stage -> Symbol) -> Maybe UnitDefn -> QuantityDict

Similar to mkQuant, but the abbreviation is moved to the third argument (Maybe String), and the Symbol is now dependent on Stage.

qw :: (Quantity q, MayHaveUnit q) => q -> QuantityDict

Smart constructor for a QuantityDict from another Quantity with units.

vc :: String -> NP -> Symbol -> Space -> QuantityDict

Creates a QuantityDict from a UID, term (NP), Symbol, and Space.

vc'' :: Idea c => c -> Symbol -> Space -> QuantityDict

Creates a QuantityDict from an Idea, Symbol, and Space.

vcSt :: String -> NP -> (Stage -> Symbol) -> Space -> QuantityDict

Similar to vc, but creates a QuantityDict from something that knows about Stages.

vcUnit :: String -> NP -> Symbol -> Space -> UnitDefn -> QuantityDict

Creates a QuantityDict from a UID, term (NP), Symbol, Space, and unit (UnitDefn).