Language.Drasil.Chunk.Relation

Description

For adding a relation (expression) to a concept.

Chunk Type

data RelationConcept

For a concept (ConceptChunk) that also has a Relation attached.

Ex. We can describe a pendulum arm and then apply an associated equation so that we know its behaviour.

Constructors

makeRC :: UID -> NP -> Sentence -> Relation -> RelationConcept

Create a RelationConcept from a given UID, term (NP), definition (Sentence), and Relation.

addRelToCC :: Concept c => c -> UID -> Relation -> RelationConcept

Create a new RelationConcept from an old Concept. Takes a Concept, new UID and relation.