Language.Drasil.Chunk.UncertainQuantity

Documentation

data UncertQ

UncertQs are conceptual symbolic quantities with constraints and an Uncertainty. Contains a ConstrConcept and an Uncertainty.

data UncertainChunk

UncertainChunk is a symbolic quantity with constraints, a typical value, and an uncertainty. Contains a ConstrainedChunk and an Uncertainty.

Constructors

UCh

Fields

_conc :: ConstrainedChunk

_unc' :: Uncertainty

uq :: (Quantity c, Constrained c, Concept c, HasReasVal c, MayHaveUnit c) => c -> Uncertainty -> UncertQ

Smart constructor that requires a Quantity, a percentage, and a typical value with an Uncertainty.

uqc :: IsUnit u => String -> NP -> String -> Symbol -> u -> Space -> [ConstraintE] -> Expr -> Uncertainty -> UncertQ

Uncertainty quantity (uq) but with a constraint.

uqcND :: IsUnit u => String -> NP -> Symbol -> u -> Space -> [ConstraintE] -> Expr -> Uncertainty -> UncertQ

Uncertainty quantity constraint (uqc) without a description.

uncrtnChunk :: (Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => c -> Uncertainty -> UncertainChunk

Smart constructor that can project to an UncertainChunk (also given an Uncertainty).

uvc :: String -> NP -> Symbol -> Space -> [ConstraintE] -> Expr -> Uncertainty -> UncertainChunk

Creates an uncertain variable chunk. Takes UID, term (NP), Symbol, Space, Constrains, Expr, and Uncertainty.

uncrtnw :: (HasUncertainty c, Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => c -> UncertainChunk

Projection function into an UncertainChunk from UncertQ or an UncertainChunk.