Language.Drasil.Chunk.UnitDefn

Description

For defining units built from a concept.

Classes

class MayHaveUnit u where

Types may contain a unit (UnitDefn).

Methods

getUnit :: u -> Maybe UnitDefn

class (Idea u, Definition u, HasUnitSymbol u) => IsUnit u where

Units are Ideas with a Definition which store a unit symbol. They must also be explicitly declared to be instances of IsUnit.

Minimal complete definition

udefn, getUnits

Methods

getUnits :: u -> [UID]

Holds units as a list of UID.

class TempHasUnit u where

Temporary class to make sure chunks have a unit (in order to eventually get rid of MayHaveUnit).

Methods

findUnit :: u -> UnitDefn

Chunk Type

data UnitDefn

For defining units. It has a ConceptChunk (that defines what kind of unit it is), a unit symbol, maybe another (when it is a synonym), perhaps a definition, and a list of UID of the units that make up the definition.

Ex. Meter is a unit of length defined by the symbol (m).

Constructors

UD

Fields

_vc :: ConceptChunk

_cas :: UnitSymbol

_cu :: [UID]

Constructors

makeDerU :: ConceptChunk -> UnitEquation -> UnitDefn

Create a derived unit chunk from a concept and a unit equation.

newUnit :: String -> UnitEquation -> UnitDefn

Smart constructor for new derived units from existing units.

derUC :: String -> String -> String -> Symbol -> UDefn -> UnitDefn

Uses self-plural term.

Create a derived unit chunk from a UID, term (String), definition, Symbol, and unit equation.

derUC' :: String -> String -> String -> Symbol -> UDefn -> UnitDefn

Uses term that pluralizes by adding "s" to the end.

Create a derived unit chunk from a UID, term (String), definition, Symbol, and unit equation.

derUC'' :: String -> NP -> String -> Symbol -> UDefn -> UnitDefn

Create a derived unit chunk from a UID, term (NP), definition, Symbol, and unit equation.

fund :: String -> String -> String -> UnitDefn

Smart constructor for a "fundamental" unit.

fund' :: String -> String -> Symbol -> UnitDefn

Variant of the fund, useful for degree.

derCUC :: String -> String -> String -> Symbol -> UnitEquation -> UnitDefn

Create a SI_Unit with two Symbol representations. The created NP is self-plural.

derCUC' :: String -> String -> String -> Symbol -> UnitEquation -> UnitDefn

Similar to derCUC, but the created NP has the AddS plural rule.

derCUC'' :: String -> NP -> String -> Symbol -> UnitEquation -> UnitDefn

Create a derived unit chunk from a UID, term (NP), definition, Symbol, and unit equation.

unitWrapper :: IsUnit u => u -> UnitDefn

For allowing lists to mix together chunks that are units by projecting them into a UnitDefn. For now, this only works on UnitDefns.

Unit Combinators (UnitEquations)

(^:) :: UnitDefn -> Integer -> UnitEquation

Combinator for raising a unit to a power.

(/:) :: UnitDefn -> UnitDefn -> UnitEquation

Combinator for dividing one unit by another.

(*:) :: UnitDefn -> UnitDefn -> UnitEquation

Combinator for multiplying two units together.

(*$) :: UnitDefn -> UnitEquation -> UnitEquation

Combinator for multiplying a unit and a symbol.

(/$) :: UnitDefn -> UnitEquation -> UnitEquation

Combinator for dividing a unit and a symbol.

(^$) :: UnitEquation -> UnitEquation -> UnitEquation

Combinator for mulitiplying two unit equations.

Unit Relation Functions

scale :: IsUnit s => Double -> s -> UDefn

Combinator for scaling one unit by some number.

shift :: IsUnit s => Double -> s -> UDefn

Combinator for shifting one unit by some number.

Helpers

fromUDefn :: UDefn -> USymb

Generates a default unit symbol.

unitCon :: String -> ConceptChunk

Helper for fundamental unit concept chunk creation. Uses the same String for the identifier, term, and definition.

getCu :: UnitEquation -> [UID]

Get a list of UID of the units that make up the UnitEquation.

compUnitDefn :: UnitDefn -> UnitDefn -> Ordering

We don't want an Ord on units, but this still allows us to compare them.