Language.Drasil.Chunk.Unital

Description

Defines chunks to add units to a quantity. Similar to UnitaryChunk.

Chunk Type

data UnitalChunk

Similar to a DefinedQuantityDict, UnitalChunks are concepts with quantities that must have a unit definition. Contains DefinedQuantityDicts and a UnitDefn.

Ex. A pendulum arm is a tangible object with a symbol (l) and units (cm, m, etc.).

Constructors

UC

Fields

_defq' :: DefinedQuantityDict

_uni :: UnitDefn

Constructors

uc :: (Concept c, IsUnit u) => c -> Symbol -> Space -> u -> UnitalChunk

Used to create a UnitalChunk from a Concept, Symbol, and Unit.

uc' :: IsUnit u => String -> NP -> Sentence -> Symbol -> Space -> u -> UnitalChunk

Similar to uc, except it builds the Concept portion of the UnitalChunk from a given UID, term, and definition (as a Sentence) which are its first three arguments.

ucStaged :: (Concept c, IsUnit u) => c -> (Stage -> Symbol) -> Space -> u -> UnitalChunk

Similar to uc, but Symbol is dependent on the Stage.

ucStaged' :: IsUnit u => String -> NP -> Sentence -> (Stage -> Symbol) -> Space -> u -> UnitalChunk

Similar to uc', but Symbol is dependent on the Stage.

ucuc :: (Quantity c, Concept c, MayHaveUnit c) => c -> UnitDefn -> UnitalChunk

Attach units to a chunk that has a symbol and definition.

ucw :: (Unitary c, Concept c, MayHaveUnit c) => c -> UnitalChunk

Constructs a UnitalChunk from a Concept with Units.