Language.Drasil.Chunk.Unitary

Description

Defines chunks to add units to a quantity. Similar to UnitalChunk.

Chunk Types

class Quantity c => Unitary c where

A Unitary is a Quantity that must have a unit.

Methods

unit :: c -> UnitDefn

data UnitaryChunk

UnitaryChunks are for ideas with quantities that must have units. Contains a QuantityDict and a UnitDefn.

Ex. A pendulum arm is an idea associated with a symbol (l) and units (cm, m, etc.).

Constructors

mkUnitary :: (Unitary u, MayHaveUnit u) => u -> UnitaryChunk

Makes a UnitaryChunk from a quantity with a unit.

unitary :: IsUnit u => String -> NP -> Symbol -> u -> Space -> UnitaryChunk

Builds the QuantityDict part from the UID, term (NP), Symbol, and Space. Assumes there's no abbreviation.

unitary' :: IsUnit u => String -> NP -> (Stage -> Symbol) -> u -> Space -> UnitaryChunk

Same as unitary but with a Symbol that changes based on the Stage.

unit_symb :: Unitary c => c -> USymb

Helper for getting the unit's Symbol from a chunk, as opposed to the symbols of the chunk itself.