Language.Drasil.Chunk.UnitaryConcept

Description

Adds concepts to a quantitative idea with units.

Chunk Type

data UnitaryConceptDict

Contains a UnitaryChunk, a definition, and a list of related UIDs.

Ex. A pendulum arm is an idea associated with a symbol (l) and units (cm, m, etc.).

Constructor

ucw :: (Unitary c, Concept c, MayHaveUnit c) => c -> UnitaryConceptDict

Constructs a UnitaryConceptDict from a Concept with Units.