Language.Drasil.Classes

Description

Defining all the classes which represent knowledge-about-knowledge

Documentation

class HasUID c => NamedIdea c where

A NamedIdea is a term that we've identified (has a UID) as being worthy of naming.

Methods

term :: Lens' c NP

Lens to the term (a noun phrase).

class HasSpace c where

HasSpace is anything which has a Space.

Methods

typ :: Lens' c Space

Provides a Lens to the Space.

class HasUnitSymbol u where

Some chunks store a unit symbol.

Methods

usymb :: u -> USymb

Provides the ability to hold a unit symbol (USymb).

class HasReference c where

A class that contains a list of References.

Methods

getReferences :: Lens' c [Reference]

Provides a Lens to the References.

class HasReasVal c where

A Quantity that could have a reasonable value.

Methods

reasVal :: Lens' c (Maybe Expr)

Provides a Lens to the possible reasonable value.

class HasDerivation c where

A class that might have a Derivation.

Methods

derivations :: Lens' c (Maybe Derivation)

Provides a Lens to a possible derivation.

class HasAdditionalNotes c where

Records any additional notes needed to avoid losing information

Methods

getNotes :: Lens' c [Sentence]

Provides a Lens to the notes.

class NamedIdea c => Idea c where

An Idea is the combination of a NamedIdea and a CommonIdea. In other words, it may have an acronym/abbreviation.

Methods

getA :: c -> Maybe String

Gets the acronym/abbreviation.

class Definition c where

Defines a chunk.

Methods

defn :: Lens' c Sentence

Provides (a Lens to) the definition for a chunk.

class ConceptDomain c where

Some concepts have a domain (related information encoded in UIDs to other chunks).

Methods

cdom :: c -> [UID]

Provides Getter for the concept domain tags for a chunk

cdom should be exported for use by the Drasil framework, but should not be exported beyond that.

class Constrained c where

The Constrained class is a Quantity that has value constraints. It does not enforce Quantity at this point.

Methods

constraints :: Lens' c [ConstraintE]

Provides a Lens to the Constraints.

class NamedIdea c => CommonIdea c where

CommonIdea is a NamedIdea with the additional constraint that it must have an abbreviation.

Methods

abrv :: c -> String

Introduces abrv which necessarily provides an abbreviation.

class DefiningExpr c where

A better version of ExprRelat that holds an Expr.

Methods

defnExpr :: Lens' c Expr

Provides a Lens to the expression.

class Display c where

Data that can convert into a Displayable Expr.

Methods

toDispExpr :: c -> DisplayExpr

class (Idea c, HasSpace c, HasSymbol c) => Quantity c

A Quantity is an Idea with a Space and a Symbol. In theory, it should also restrict to being a part of MayHaveUnit, but that causes all sorts of import cycles (or lots of orphans).

class HasUncertainty c where

HasUncertainty is just a chunk with some uncertainty associated to it. This uncertainty is represented as a decimal value between 0 and 1 (percentage).

Methods

unc :: Lens' c Uncertainty

Provides the Lens to an Uncertainty.

type Concept c = (Idea c, Definition c, ConceptDomain c)

Concepts are Ideas with definitions and domains.

class HasSymbol c => Callable c

Some chunks can be called like functions.

class HasSymbol c => IsArgumentName c

Members must have a named argument.

class (Idea u, Definition u, HasUnitSymbol u) => IsUnit u where

Units are Ideas with a Definition which store a unit symbol. They must also be explicitly declared to be instances of IsUnit.

Methods

udefn :: u -> Maybe UDefn

May have a unit definition.

getUnits :: u -> [UID]

Holds units as a list of UID.

class UnitEq u where

Contains a UDefn

Methods

uniteq :: Lens' u UDefn

Provides the Lens to a unit definition.