Language.Drasil.Classes.Core

Description

Defining the core classes which represent knowledge-about-knowledge

Documentation

class HasUID c where

The most basic item: having a unique identifier key, here a UID.

Methods

uid :: Lens' c UID

Provides a unique id for internal Drasil use.

class HasShortName s where

A ShortName is the text to be displayed for a link. Used for referencing within a document that can include symbols and whatnot if required. Visible in the typeset documents (pdf).

Methods

shortname :: s -> ShortName

class HasRefAddress b where

Members must have a reference address.

Methods

getRefAdd :: b -> String

Provides the ability to hold a reference address.

class HasSymbol c where

A HasSymbol is anything which has a Symbol.

Methods

symbol :: c -> Stage -> Symbol

Provides the Symbol for a particular stage of generation.