Language.Drasil.Constraint

Description

Defines types and functions for constrained values.

Types

data Constraint a where

Holds constraints. May occur between an interval of Expr, a list of Doubles, or a list of Strings.

Constructors

Range :: ConstraintReason -> RealInterval a a -> Constraint a
By default, physical and software constraints are ranges.

Elem :: ConstraintReason -> a -> Constraint a

type ConstraintE = Constraint Expr

Type synonym for ConstraintE

data ConstraintReason

The reason behind the constraint's existence.

Constructors

Physical

Software

Functions

physRange :: RealInterval Expr Expr -> ConstraintE

Smart constructor for range of Physical constraints between two given expressions.

sfwrRange :: RealInterval Expr Expr -> ConstraintE

Smart constructor for range of Software constraints between two given expressions.

physElem :: Expr -> ConstraintE

sfwrElem :: Expr -> ConstraintE

isPhysC :: Constraint e -> Bool

Helpful for filtering for Physical constraints. True if constraint is Physical.

isSfwrC :: Constraint e -> Bool

Helpful for filtering for Software constraints. True if constraint is Software.