Language.Drasil.Data.Citation

Description

Contains all necessary types and constructors for citing sources in Drasil.

Types

data CiteField

Fields used in citations.

Constructors

Address String

Author People

BookTitle String
Used for InCollection references only.

Chapter Int

Edition Int

Editor People

HowPublished HP
Can be published via URL or something else.

Institution String

Journal String

Month Month

Note String

Number Int

Organization String

Pages [Int]
Range of pages (ex1. 1-32; ex2. 7,31,52-55).

Publisher String

School String

Series String

Title String

Type String
BibTeX "type" field.

Volume Int

Year Int

data HP

How something is published. Necessary for URLs to work properly.

Constructors

URL String

Verb String

data CitationKind

External references come in many flavours. Articles, Books, etc. (we are using the types available in Bibtex).

Constructors

Article

Book

Booklet

InBook

InCollection

InProceedings

Manual

MThesis

Misc

PhDThesis

Proceedings

TechReport

Unpublished

Class

class HasFields c where

Citations should have a fields (CiteField).

Methods

getFields :: Lens' c [CiteField]

Provides a Lens to CiteFields.

CiteField Constructors

People -> CiteField

author :: People -> CiteField

Smart field constructor for a CiteField.

editor :: People -> CiteField

Smart field constructor for a CiteField.

String -> CiteField

address :: String -> CiteField

Smart field constructor for a CiteField.

bookTitle :: String -> CiteField

Smart field constructor for a CiteField.

howPublished :: String -> CiteField

Smart field constructor for a CiteField.

howPublishedU :: String -> CiteField

URL version of howPublished.

Smart field constructor for a CiteField.

institution :: String -> CiteField

Smart field constructor for a CiteField.

journal :: String -> CiteField

Smart field constructor for a CiteField.

note :: String -> CiteField

Smart field constructor for a CiteField.

organization :: String -> CiteField

Smart field constructor for a CiteField.

publisher :: String -> CiteField

Smart field constructor for a CiteField.

school :: String -> CiteField

Smart field constructor for a CiteField.

series :: String -> CiteField

Smart field constructor for a CiteField.

title :: String -> CiteField

Smart field constructor for a CiteField.

typeField :: String -> CiteField

Smart field constructor for a CiteField.

Int -> CiteField

chapter :: Int -> CiteField

Smart field constructor for a CiteField.

edition :: Int -> CiteField

Smart field constructor for a CiteField.

number :: Int -> CiteField

Smart field constructor for a CiteField.

volume :: Int -> CiteField

Smart field constructor for a CiteField.

year :: Int -> CiteField

Smart field constructor for a CiteField.

[Int] -> CiteField

pages :: [Int] -> CiteField

Smart field constructor for a CiteField.

CiteField -> CiteField

month :: Month -> CiteField

Smart field constructor for a CiteField.

Comparators

compareAuthYearTitle :: HasFields c => c -> c -> Ordering

Orders two authors. If given two of the exact same authors, year, and title, returns an error.