Language.Drasil.Data.Date

Description

Custom type for dates, in this case Month. This should, in time, be switched out in favour of a proper package designed to handle all the complexities of dates.

Type

data Month

Custom type for months (abbreviated).

Constructors

Jan

Feb

Mar

Apr

May

Jun

Jul

Aug

Sep

Oct

Nov

Dec