Language.Drasil.DecoratedReference

Description

References that have extra information.

Type

data DecRef

For holding a Reference that is decorated with extra information (ex. page numbers, equation sources, etc.).

Constructors

DR

Fields

_rf :: Reference

refInfo :: RefInfo

Class

class HasDecRef c where

A class that contains a list of decorated references (DecRefs).

Methods

getDecRefs :: Lens' c [DecRef]

Provides a Lens to the DecRefs.

Constructors

dRef :: (HasUID r, HasRefAddress r, HasShortName r) => r -> DecRef

Same as ref, but for DecRef instead of Reference.

dRefInfo :: (HasUID r, HasRefAddress r, HasShortName r) => r -> RefInfo -> DecRef

For creating a decorated reference (DecRef) with extra reference information (RefInfo).