Language.Drasil.Development

Documentation

class NounPhrase n where

Minimal complete definition

phraseNP, pluralNP, sentenceCase, titleCase

Methods

phraseNP :: n -> Sentence

pluralNP :: n -> PluralForm

dep :: Expr -> [String]

Get dependencies from an equation

names :: Expr -> [String]

Generic traverse of all positions that could lead to names

names' :: Expr -> [String]

Generic traverse of all positions that could lead to names, without functions. FIXME : this should really be done via post-facto filtering, but right now the information needed to do this is not available!

sdep :: Sentence -> [UID]

lnames :: Sentence -> [UID]

Generic traverse of all positions that could lead to reference UID from sentences

lnames' :: [Sentence] -> [UID]

precA :: ArithOper -> Int

prec - precedence for Binary-Associative (Commutative) operators

precB :: BoolOper -> Int

eprec :: Expr -> Int

eprec - Expression precedence