Language.Drasil.Development.Sentence

Description

Various helpers for building Sentences from other bits. Really ought to be moved out to (likely) docLang, but is here for now.

All Lowercase

phrase :: NamedIdea n => n -> Sentence

Helper for getting the phrase from a NamedIdea using it's UID.

plural :: NamedIdea n => n -> Sentence

Helper for getting the plural of a phrase from a NamedIdea.

phrasePoss :: NamedIdea n => n -> Sentence

Singular possesive function

Helper for getting the possesive cases from the term of a NamedIdea.

pluralPoss :: NamedIdea n => n -> Sentence

Plural possesive function

Helper for getting the possesive cases from the term of a NamedIdea.

Capitalize First Word

atStart :: NamedIdea n => n -> Sentence

Singular sentence case.

Helper function for getting the sentence case of a noun phrase from a NamedIdea.

atStart' :: NamedIdea n => n -> Sentence

Plural sentence case.

Helper function for getting the sentence case of a noun phrase from a NamedIdea.

Capitalize All Words

titleize :: NamedIdea n => n -> Sentence

Singular title case.

Helper function for getting the title case of a noun phrase from a NamedIdea.

titleize' :: NamedIdea n => n -> Sentence

Plural title case.

Helper function for getting the title case of a noun phrase from a NamedIdea.

Short Form (lowercase)

short :: Idea c => c -> Sentence

Get short form (if it exists), else get term of an Idea. Uses the UID of the Idea in a Ch Sentence constructor to get the short form using getA. getA may return Nothing, in which case lookupS uses the term, where lookupS is the main helper for looking up the short form of a Ch Sentence.

introduceAbb :: Idea n => n -> Sentence

Helper for common pattern of introducing the title-case version of a noun phrase (from an Idea) followed by its abbreviation in parentheses.