Language.Drasil.DisplayExpr

Documentation

data DisplayExpr where

A variant of the expression language that allows for extensions of Expr that wouldn't normally type check, or that we only care for displaying.

Constructors

AlgebraicExpr :: Expr -> DisplayExpr

SpaceExpr :: Space -> DisplayExpr

BinOp :: DisplayBinOp -> DisplayExpr -> DisplayExpr -> DisplayExpr

AssocBinOp :: DisplayAssocBinOp -> [DisplayExpr] -> DisplayExpr

data DisplayBinOp

Binary Display Operations.

Constructors

Defines

IsIn

data DisplayAssocBinOp

Associative Binary Display Operations.

Constructors

DAnd

Equivalence