Language.Drasil.Document

Description

Document Description Language

Documentation

data SecCons

Section Contents are split into subsections or contents, where contents are standard layout objects (see Contents).

Constructors

Sub Section

Con Contents

data Partition

Constructors

Sections

Part

Chapter

data Section

Sections have a title (Sentence), a list of contents (SecCons) and a shortname (Reference).

Constructors

Section

Fields

tle :: Title

cons :: [SecCons]

_lab :: Reference

lab :: Lens' Section Reference

data Document

A Document has a Title (Sentence), Author(s) (Sentence), and Sections which hold the contents of the document.

Constructors

Document Title Author ShowTableOfContents [Section]

data ShowTableOfContents

Determines whether or not the table of contents appears on the generated artifacts.

Constructors

ToC

NoToC

checkToC :: Document -> Document

Manually removes the first section of a document (table of contents section).

llcc :: Reference -> RawContent -> LabelledContent

Smart constructor for labelled content chunks.

ulcc :: RawContent -> UnlabelledContent

Smart constructor for unlabelled content chunks (no Reference).

mkParagraph :: Sentence -> Contents

Smart constructor that wraps UnlabelledContent into Contents.

mkFig :: Reference -> RawContent -> Contents

Smart constructor that wraps LabelledContent into Contents.

mkRawLC :: RawContent -> Reference -> LabelledContent

Smart constructor similar to llcc, but takes in RawContent first.

section :: Sentence -> [Contents] -> [Section] -> Reference -> Section

Smart constructor for creating Sections with a title (Sentence), introductory contents (ie. paragraphs, tables, etc.), a list of subsections, and a shortname (Reference).

extractSection :: Document -> [Section]

Smart constructor for retrieving the contents (Sections) from a Document.

getSec :: Section -> [Section]

Smart constructor for retrieving the subsections (Sections) within a Section.

getSecCons :: SecCons -> [Section]

Helper to retrieve subsections (Sections) from section contents (SecCons).

fig :: Lbl -> Filepath -> RawContent

Figure smart constructor with a Lbl and a Filepath. Assumes 100% of page width as max width.

figWithWidth :: Lbl -> Filepath -> MaxWidthPercent -> RawContent

Figure smart constructor that allows for customized max widths.

makeTabRef :: String -> Reference

Create a reference for a table. Takes in the name of a table (which will also be used for its shortname).

makeFigRef :: String -> Reference

Create a reference for a figure. Takes in the name of a figure (which will also be used for its shortname).

makeSecRef :: String -> Sentence -> Reference

Create a reference for a section. Takes in the name of a section and a shortname for the section.

makeURI :: UID -> String -> ShortName -> Reference

Create a reference for a URI.