Language.Drasil.Document.Core

Description

Contains types and functions common to aspects of generating documents.

Lists

data ListType

Denotes the different possible types that can be used as a list.

Constructors

Bullet [(ItemType, Maybe String)]
Bulleted list.

Numeric [(ItemType, Maybe String)]
Enumerated list.

Simple [ListTuple]
Simple list with items denoted by :. Renders as "Title: Item"

Desc [ListTuple]
Descriptive list, renders as "Title: Item" (see ListTuple).

Definitions [ListTuple]
Renders a list of "Title is the Item".

data ItemType

Denotes how something should behave in a list (ListType).

Constructors

Flat Sentence
Standard singular item.

Nested Header ListType
Nest a list (ListType) as an item.

type MaxWidthPercent = Float

MaxWidthPercent should be kept in the range 1-100. Values outside this range may have unexpected results. Used for specifying max figure width as pagewidth*MaxWidthPercent/100.

type Title = Sentence

type Author = Sentence

type Header

Arguments

= Sentence
Used when creating sublists.

type Depth = Int

type Width = Float

type Height = Float

type ListTuple

Arguments

= (Title, ItemType, Maybe String)
Formats as Title: Item. For use in lists.

type Filepath = String

type Lbl

Arguments

= Sentence
Label.

Contents

data Contents

Contents may be labelled or unlabelled.

Constructors

UlC UnlabelledContent

LlC LabelledContent

data DType

Types of definitions (general, instance, theory, or data).

Constructors

General

Instance

Theory

Data

data HasCaption

Indicates whether a figure has a caption or not.

Constructors

NoCaption

WithCaption

data RawContent

Types of layout objects we deal with explicitly.

Constructors

Table [Sentence] [[Sentence]] Title Bool
table has: header-row, data(rows), label/caption, and a bool that determines whether or not to show label.

Paragraph Sentence
Paragraphs are just sentences.

EqnBlock ModelExpr
Block of Equations holds an expression.

DerivBlock Sentence [RawContent]
Grants the ability to label a group of RawContent.

Enumeration ListType
For enumerated lists.

Defini DType [(Identifier, [Contents])]
Defines something with a type, identifier, and Contents.

Figure Lbl Filepath MaxWidthPercent HasCaption
For creating figures in a document includes whether the figure has a caption.

Bib BibRef
Grants the ability to reference something.

Graph [(Sentence, Sentence)] (Maybe Width) (Maybe Height) Lbl
Contain a graph with coordinates (Sentences), maybe a width and height, and a label (Sentence).

CodeBlock CodeExpr
Block for codes TODO: Fill this one in.

type Identifier = String

An identifier is just a String.

data LabelledContent

Contains a Reference and RawContent.

Constructors

LblC

Fields

_ref :: Reference

_ctype :: RawContent

newtype UnlabelledContent

Only contains RawContent.

Constructors

UnlblC

Fields

_cntnts :: RawContent

ref :: Lens' LabelledContent Reference

ctype :: Lens' LabelledContent RawContent

cntnts :: Iso' UnlabelledContent RawContent

class HasContents c where

Members of this class must have RawContent.

Methods

accessContents :: Lens' c RawContent

Provides a Lens to the RawContent.

Helper

prependLabel :: RawContent -> IRefProg

Helper to prepend labels to LabelledContent when referencing.