Language.Drasil.Expr

Description

The Drasil Expression language

Expression Types

type Relation = Expr

A relation is just an expression (Expr).

data ArithBinOp

Arithmetic operators (fractional, power, and subtraction).

Constructors

Frac

Pow

Subt

data EqBinOp

Equality operators (equal or not equal).

Constructors

Eq

NEq

data BoolBinOp

Conditional and Biconditional operators (Expressions can imply one another, or exist if and only if another expression exists).

Constructors

Impl

Iff

data LABinOp

Index operator.

Constructors

Index

data OrdBinOp

Ordered binary operators (less than, greater than, less than or equal to, greater than or equal to).

Constructors

Lt

Gt

LEq

GEq

data VVVBinOp

Vector x Vector -> Vector binary operations (cross product).

Constructors

Cross

data VVNBinOp

Vector x Vector -> Number binary operations (dot product).

Constructors

Dot

data AssocArithOper

Associative operators (adding and multiplication). Also specifies whether it is for integers or for real numbers.

Constructors

AddI

AddRe

MulI

MulRe

data AssocBoolOper

Associative boolean operators (and, or).

Constructors

And

Or

data UFunc

Unary functions (abs, log, ln, sin, etc.).

Constructors

Abs

Log

Ln

Sin

Cos

Tan

Sec

Csc

Cot

Arcsin

Arccos

Arctan

Exp

Sqrt

Neg

data UFuncB

Bool -> Bool operators.

Constructors

Not

data UFuncVV

Vector -> Vector operators.

Constructors

NegV

data UFuncVN

Vector -> Number operators.

Constructors

Norm

Dim

data Completeness

For case expressions (either complete or incomplete).

Constructors

Complete

Incomplete

Expr

data Expr where

Drasil expressions.

Constructors

Dbl :: Double -> Expr
Turns a decimal value (Double) into an expression.

Int :: Integer -> Expr
Turns an integer into an expression.

ExactDbl :: Integer -> Expr
Represents decimal values that are exact as integers.

Str :: String -> Expr
Turns a string into an expression.

Perc :: Integer -> Integer -> Expr
Turns two integers into a fraction (or percent).

AssocA :: AssocArithOper -> [Expr] -> Expr
Takes an associative arithmetic operator with a list of expressions.

AssocB :: AssocBoolOper -> [Expr] -> Expr
Takes an associative boolean operator with a list of expressions.

Deriv :: DerivType -> Expr -> UID -> Expr
Derivative syntax is: Type (Partial or Total) -> principal part of change -> with respect to For example: Deriv Part y x1 would be (dy/dx1).

C :: UID -> Expr
C stands for Chunk, for referring to a chunk in an expression. Implicitly assumes that the chunk has a symbol.

FCall :: UID -> [Expr] -> [(UID, Expr)] -> Expr
A function call accepts a list of parameters and a list of named parameters. For example

F(x) is (FCall F [x] []).

F(x,y) would be (FCall F [x,y]).

F(x,n=y) would be (FCall F [x] [(n,y)]).

Case :: Completeness -> [(Expr, Relation)] -> Expr
For multi-case expressions, each pair represents one case.

Matrix :: [[Expr]] -> Expr
Represents a matrix of expressions.

UnaryOp :: UFunc -> Expr -> Expr
Unary operation for most functions (eg. sin, cos, log, etc.).

UnaryOpB :: UFuncB -> Expr -> Expr
Unary operation for Bool -> Bool operations.

UnaryOpVV :: UFuncVV -> Expr -> Expr
Unary operation for Vector -> Vector operations.

UnaryOpVN :: UFuncVN -> Expr -> Expr
Unary operation for Vector -> Number operations.

ArithBinaryOp :: ArithBinOp -> Expr -> Expr -> Expr
Binary operator for arithmetic between expressions (fractional, power, and subtraction).

BoolBinaryOp :: BoolBinOp -> Expr -> Expr -> Expr
Binary operator for boolean operators (implies, iff).

EqBinaryOp :: EqBinOp -> Expr -> Expr -> Expr
Binary operator for equality between expressions.

LABinaryOp :: LABinOp -> Expr -> Expr -> Expr
Binary operator for indexing two expressions.

OrdBinaryOp :: OrdBinOp -> Expr -> Expr -> Expr
Binary operator for ordering expressions (less than, greater than, etc.).

VVVBinaryOp :: VVVBinOp -> Expr -> Expr -> Expr
Binary operator for Vector x Vector -> Vector operations (cross product).

VVNBinaryOp :: VVNBinOp -> Expr -> Expr -> Expr
Binary operator for Vector x Vector -> Number operations (dot product).

Operator :: AssocArithOper -> DomainDesc Expr Expr -> Expr -> Expr
Operators are generalized arithmetic operators over a DomainDesc of an Expr. Could be called BigOp. ex: Summation is represented via Add over a discrete domain.

RealI :: UID -> RealInterval Expr Expr -> Expr
A different kind of IsIn. A UID is an element of an interval.

Binary Operators

($=) :: Expr -> Expr -> Expr infixr 4

Smart constructor for equating two expressions.

($!=) :: Expr -> Expr -> Expr

Smart constructor for showing that two expressions are not equal.

($<) :: Expr -> Expr -> Expr

Less than.

Smart constructor for ordering two equations.

($>) :: Expr -> Expr -> Expr

Greater than.

Smart constructor for ordering two equations.

($<=) :: Expr -> Expr -> Expr

Less than or equal to.

Smart constructor for ordering two equations.

($>=) :: Expr -> Expr -> Expr

Greater than or equal to.

Smart constructor for ordering two equations.

($.) :: Expr -> Expr -> Expr

Smart constructor for the dot product of two equations.

addI :: Expr -> Expr -> Expr

Add two expressions (Integers).

addRe :: Expr -> Expr -> Expr

Add two expressions (Real numbers).

mulI :: Expr -> Expr -> Expr

Multiply two expressions (Integers).

mulRe :: Expr -> Expr -> Expr

Multiply two expressions (Real numbers).

($-) :: Expr -> Expr -> Expr

Smart constructor for subtracting two expressions.

($/) :: Expr -> Expr -> Expr infixl 7

Smart constructor for dividing two expressions.

($^) :: Expr -> Expr -> Expr infixr 8

Smart constructor for rasing the first expression to the power of the second.

($=>) :: Expr -> Expr -> Expr

Smart constructor to show that one expression implies the other (conditional operator).

($<=>) :: Expr -> Expr -> Expr

Smart constructor to show that an expression exists if and only if another expression exists (biconditional operator).

($&&) :: Expr -> Expr -> Expr infixr 9

Smart constructor for the boolean and operator.

($||) :: Expr -> Expr -> Expr infixr 9

Smart constructor for the boolean or operator.

Misc.

type Variable = String

The variable type is just a renamed String.

data DerivType

Determines the type of the derivative (either a partial derivative or a total derivative).

Constructors

Part

Total