Language.Drasil.Expr.Class

Documentation

class ExprC r where

Methods

lit :: Literal -> r

($=) :: r -> r -> r infixr 4

($!=) :: r -> r -> r

($<) :: r -> r -> r

Smart constructor for ordering two equations.

($>) :: r -> r -> r

Smart constructor for ordering two equations.

($<=) :: r -> r -> r

Smart constructor for ordering two equations.

($>=) :: r -> r -> r

Smart constructor for ordering two equations.

($.) :: r -> r -> r

Smart constructor for the dot product of two equations.

($+) :: r -> r -> r

Add two expressions.

($*) :: r -> r -> r

Multiply two expressions.

($-) :: r -> r -> r

($/) :: r -> r -> r infixl 7

($^) :: r -> r -> r infixr 8

($=>) :: r -> r -> r

($<=>) :: r -> r -> r

($&&) :: r -> r -> r infixr 9

($||) :: r -> r -> r infixr 9

in' :: r -> r -> r

Smart constructor for set-theoretic membership relation. Added ' to avoid conflict.

abs_ :: r -> r

Smart constructor for taking the absolute value of an expression.

neg :: r -> r

Smart constructor for negating an expression.

log :: r -> r

Smart constructor to take the log of an expression.

ln :: r -> r

Smart constructor to take the ln of an expression.

sqrt :: r -> r

Smart constructor to take the square root of an expression.

sin :: r -> r

Smart constructor to apply sin to an expression.

cos :: r -> r

Smart constructor to apply cos to an expression.

tan :: r -> r

Smart constructor to apply tan to an expression.

sec :: r -> r

Smart constructor to apply sec to an expression.

csc :: r -> r

Smart constructor to apply csc to an expression.

cot :: r -> r

Smart constructor to apply cot to an expression.

arcsin :: r -> r

Smart constructor to apply arcsin to an expression.

arccos :: r -> r

Smart constructor to apply arccos to an expression.

arctan :: r -> r

Smart constructor to apply arctan to an expression.

exp :: r -> r

Smart constructor for the exponential (base e) function.

dim :: r -> r

Smart constructor for calculating the dimension of a vector.

norm :: r -> r

Smart constructor for calculating the normal form of a vector.

negVec :: r -> r

Smart constructor for negating vectors.

not_ :: r -> r

Smart constructor for applying logical negation to an expression.

idx :: r -> r -> r

Smart constructor for indexing.

idxOf :: r -> r -> r

Smart constructor for indexOf. Finds the index of the first occurrence of a value in a list.

defint :: Symbol -> r -> r -> r -> r

Smart constructor for the summation, product, and integral functions over an interval.

defsum :: Symbol -> r -> r -> r -> r

Smart constructor for the summation, product, and integral functions over an interval.

defprod :: Symbol -> r -> r -> r -> r

Smart constructor for the summation, product, and integral functions over an interval.

realInterval :: HasUID c => c -> RealInterval r r -> r

Smart constructor for 'real interval' membership.

euclidean :: [r] -> r

Euclidean function : takes a vector and returns the sqrt of the sum-of-squares.

cross :: r -> r -> r

Smart constructor to cross product two expressions.

vScale :: r -> r -> r

Smart constructor for vector scaling

vAdd :: r -> r -> r

Vector Addition

vSub :: r -> r -> r

Vector Subtraction

completeCase :: [(r, r)] -> r

Smart constructor for case statements with a complete set of cases.

incompleteCase :: [(r, r)] -> r

Smart constructor for case statements with an incomplete set of cases.

matrix :: [[r]] -> r

Create a matrix.

set' :: Space -> [r] -> r

Create a Set.

apply :: (HasUID f, HasSymbol f) => f -> [r] -> r

Applies a given function with a list of parameters.

sy :: (HasUID c, HasSymbol c) => c -> r

Create an Expr from a Symbolic Chunk.

frac :: (ExprC r, LiteralC r) => Integer -> Integer -> r

Smart constructor for fractions.

recip_ :: (ExprC r, LiteralC r) => r -> r

Smart constructor for rational expressions (only in 1/x form).

square :: (ExprC r, LiteralC r) => r -> r

Smart constructor to square a function.

half :: (ExprC r, LiteralC r) => r -> r

Smart constructor to half a function exactly.

oneHalf :: (ExprC r, LiteralC r) => r

1/2, as an expression.

oneThird :: (ExprC r, LiteralC r) => r

1/3rd, as an expression.

apply1 :: (ExprC r, HasUID f, HasSymbol f, HasUID a, HasSymbol a) => f -> a -> r

Similar to apply, but converts second argument into Symbols.

apply2 :: (ExprC r, HasUID f, HasSymbol f, HasUID a, HasSymbol a, HasUID b, HasSymbol b) => f -> a -> b -> r

Similar to apply, but the applied function takes two parameters (which are both Symbols).

m2x2 :: ExprC r => r -> r -> r -> r -> r

Create a two-by-two matrix from four given values. For example:

>>> m2x2 1 2 3 4
[ [1,2],
  [3,4] ]

vec2D :: ExprC r => r -> r -> r

Create a 2D vector (a matrix with two rows, one column). First argument is placed above the second.

dgnl2x2 :: (ExprC r, LiteralC r) => r -> r -> r

Creates a diagonal two-by-two matrix. For example:

>>> dgnl2x2 1 2
[ [1, 0],
  [0, 2] ]

rowVec :: ExprC r => [r] -> r

Create a row vector

columnVec :: ExprC r => [r] -> r

Create a column vector

mkSet :: ExprC r => Space -> [r] -> r