Language.Drasil.Label.Type

Description

Even though we do not have Labels per se, here we define the different ways of construction ways to mark labels.

Types

data LblType

Applying different pieces of information for a reference. An RP is a decorated internal reference. Citation is a citation. URI is for URLs and other external links.

Constructors

RP IRefProg String

Citation String

URI String

data IRefProg

Created for different forms of references. Used in LblType.

Constructors

Deferred UID
Deferred lookup; done later. Used for domains in a ConceptInstance.

RS String
Lifts a String into a RefProg.

RConcat IRefProg IRefProg
Concatenates with two subprograms.

Name
The Symbol to insert the ShortName directly.

LblType accessor

getAdd :: LblType -> String

Retrieves the String contained in a LblType.

IRefProg constructors

name :: IRefProg

Constructor that gets the Name of an IRefProg.

(+::+) :: IRefProg -> IRefProg -> IRefProg

Constructor that concatenates two subprograms.

raw :: String -> IRefProg

Constructor for a String into an IRefProg.

defer :: UID -> IRefProg

Constructor to defer a UID lookup; done later.

prepend :: String -> IRefProg

Prepends a String to an IRefProg.