Language.Drasil.ModelExpr.Class

Documentation

assocCreate :: AssocBoolOper -> [ModelExpr] -> ModelExpr

Helper for creating new smart constructors for Associative Binary operations that require at least 1 expression.

assocSanitize :: AssocBoolOper -> [ModelExpr] -> [ModelExpr]

Helper for associative operations, removes embedded variants of the same kind

class ModelExprC r where

Methods

deriv :: (HasUID c, HasSymbol c) => r -> c -> r

Gets the derivative of an ModelExpr with respect to a Symbol.

pderiv :: (HasUID c, HasSymbol c) => r -> c -> r

Gets the derivative of an ModelExpr with respect to a Symbol.

nthderiv :: (HasUID c, HasSymbol c) => Integer -> r -> c -> r

Gets the nthderivative of an ModelExpr with respect to a Symbol.

nthpderiv :: (HasUID c, HasSymbol c) => Integer -> r -> c -> r

Gets the nthderivative of an ModelExpr with respect to a Symbol.

defines :: r -> r -> r

One expression is "defined" by another.

space :: Space -> r

Space literals.

isIn :: r -> Space -> r

Check if a value belongs to a Space.

equiv :: [r] -> r

Binary associative Equivalence.

intAll :: Symbol -> r -> r

Smart constructor for the summation, product, and integral functions over all Real numbers.

sumAll :: Symbol -> r -> r

Smart constructor for the summation, product, and integral functions over all Real numbers.

prodAll :: Symbol -> r -> r

Smart constructor for the summation, product, and integral functions over all Real numbers.