Language.Drasil.ModelExpr.Development

Description

Re-export functions and types related to ModelExpr

Types

data ModelExpr where

Drasil expressions.

Constructors

Dbl :: Double -> ModelExpr
Turns a decimal value (Double) into an expression.

Int :: Integer -> ModelExpr
Turns an integer into an expression.

ExactDbl :: Integer -> ModelExpr
Represents decimal values that are exact as integers.

Str :: String -> ModelExpr
Turns a string into an expression.

Spc :: Space -> ModelExpr
Introduce Space values into the expression language.

Perc :: Integer -> Integer -> ModelExpr
Turns two integers into a fraction (or percent).

AssocA :: AssocArithOper -> [ModelExpr] -> ModelExpr
Takes an associative arithmetic operator with a list of expressions.

AssocB :: AssocBoolOper -> [ModelExpr] -> ModelExpr
Takes an associative boolean operator with a list of expressions.

Deriv :: DerivType -> ModelExpr -> UID -> ModelExpr
Derivative syntax is: Type (Partial or Total) -> principal part of change -> with respect to For example: Deriv Part y x1 would be (dy/dx1).

C :: UID -> ModelExpr
C stands for Chunk, for referring to a chunk in an expression. Implicitly assumes that the chunk has a symbol.

FCall :: UID -> [ModelExpr] -> [(UID, ModelExpr)] -> ModelExpr
A function call accepts a list of parameters and a list of named parameters. For example

F(x) is (FCall F [x] []).

F(x,y) would be (FCall F [x,y]).

F(x,n=y) would be (FCall F [x] [(n,y)]).

Case :: Completeness -> [(ModelExpr, ModelExpr)] -> ModelExpr
For multi-case expressions, each pair represents one case.

Matrix :: [[ModelExpr]] -> ModelExpr
Represents a matrix of expressions.

UnaryOp :: UFunc -> ModelExpr -> ModelExpr
Unary operation for most functions (eg. sin, cos, log, etc.).

UnaryOpB :: UFuncB -> ModelExpr -> ModelExpr
Unary operation for Bool -> Bool operations.

UnaryOpVV :: UFuncVV -> ModelExpr -> ModelExpr
Unary operation for Vector -> Vector operations.

UnaryOpVN :: UFuncVN -> ModelExpr -> ModelExpr
Unary operation for Vector -> Number operations.

ArithBinaryOp :: ArithBinOp -> ModelExpr -> ModelExpr -> ModelExpr
Binary operator for arithmetic between expressions (fractional, power, and subtraction).

BoolBinaryOp :: BoolBinOp -> ModelExpr -> ModelExpr -> ModelExpr
Binary operator for boolean operators (implies, iff).

EqBinaryOp :: EqBinOp -> ModelExpr -> ModelExpr -> ModelExpr
Binary operator for equality between expressions.

LABinaryOp :: LABinOp -> ModelExpr -> ModelExpr -> ModelExpr
Binary operator for indexing two expressions.

OrdBinaryOp :: OrdBinOp -> ModelExpr -> ModelExpr -> ModelExpr
Binary operator for ordering expressions (less than, greater than, etc.).

SpaceBinaryOp :: SpaceBinOp -> ModelExpr -> ModelExpr -> ModelExpr
Space-related binary operations.

StatBinaryOp :: StatBinOp -> ModelExpr -> ModelExpr -> ModelExpr
Statement-related binary operations.

VVVBinaryOp :: VVVBinOp -> ModelExpr -> ModelExpr -> ModelExpr
Binary operator for Vector x Vector -> Vector operations (cross product).

VVNBinaryOp :: VVNBinOp -> ModelExpr -> ModelExpr -> ModelExpr
Binary operator for Vector x Vector -> Number operations (dot product).

Operator :: AssocArithOper -> DomainDesc ModelExpr ModelExpr -> ModelExpr -> ModelExpr
Operators are generalized arithmetic operators over a DomainDesc of an Expr. Could be called BigOp. ex: Summation is represented via Add over a discrete domain.

RealI :: UID -> RealInterval ModelExpr ModelExpr -> ModelExpr
A different kind of IsIn. A UID is an element of an interval.

ForAll :: UID -> Space -> ModelExpr -> ModelExpr
Universal quantification

data UFunc

Unary functions (abs, log, ln, sin, etc.).

Constructors

Abs

Log

Ln

Sin

Cos

Tan

Sec

Csc

Cot

Arcsin

Arccos

Arctan

Exp

Sqrt

Neg

data UFuncB

Bool -> Bool operators.

Constructors

Not

data UFuncVV

Vector -> Vector operators.

Constructors

NegV

data UFuncVN

Vector -> Number operators.

Constructors

Norm

Dim

data ArithBinOp

Arithmetic operators (fractional, power, and subtraction).

Constructors

Frac

Pow

Subt

data BoolBinOp

Conditional and Biconditional operators (Expressions can imply one another, or exist if and only if another expression exists).

Constructors

Impl

Iff

data EqBinOp

Equality operators (equal or not equal).

Constructors

Eq

NEq

data LABinOp

Index operator.

Constructors

Index

data OrdBinOp

Ordered binary operators (less than, greater than, less than or equal to, greater than or equal to).

Constructors

Lt

Gt

LEq

GEq

data SpaceBinOp

Value -> Space -> Bool operators.

Constructors

IsIn

data StatBinOp

Statements involving 2 arguments.

Constructors

Defines

data VVVBinOp

Vector x Vector -> Vector binary operations (cross product).

Constructors

Cross

data VVNBinOp

Vector x Vector -> Number binary operations (dot product).

Constructors

Dot

data AssocArithOper

Associative operators (adding and multiplication). Also specifies whether it is for integers or for real numbers.

Constructors

AddI

AddRe

MulI

MulRe

data AssocBoolOper

Associative boolean operators (and, or).

Constructors

And

Or

Equivalence

data DerivType

Determines the type of the derivative (either a partial derivative or a total derivative).

Constructors

Part

Total

data Completeness

For case expressions (either complete or incomplete).

Constructors

Complete

Incomplete

Functions

expr :: Expr -> ModelExpr

meDep :: ModelExpr -> [String]

Get dependencies from an equation.

mePrec :: ModelExpr -> Int

eprec - ModelExpr precedence.

precB :: AssocBoolOper -> Int

precB - precedence for boolean-related Binary-Associative (Commutative) operators.

precA :: AssocArithOper -> Int

precA - precedence for arithmetic-related Binary-Associative (Commutative) operators.