Language.Drasil.ModelExpr.Extract

Description

Defines functions to find Chunk UIDs within ModelExprs.

Documentation

meNames :: ModelExpr -> [UID]

Generic traverse of all expressions that could lead to names.

meNamesRI :: RealInterval ModelExpr ModelExpr -> [UID]

Generic traversal of everything that could come from an interval to names (similar to meNames).

meNames' :: ModelExpr -> [UID]

Generic traverse of all positions that could lead to meNames without functions. FIXME : this should really be done via post-facto filtering, but right now the information needed to do this is not available!

meNamesRI' :: RealInterval ModelExpr ModelExpr -> [UID]

Generic traversal of everything that could come from an interval to names without functions (similar to meNames').

meDep :: ModelExpr -> [UID]

Get dependencies from an equation.