Language.Drasil.ModelExpr.Math

Description

Defines mathematical operations for the model expression language.

Documentation

abs_ :: ModelExpr -> ModelExpr

Smart constructor for taking the absolute value of an expression.

neg :: ModelExpr -> ModelExpr

Smart constructor for negating an expression.

log :: ModelExpr -> ModelExpr

Smart constructor to take the log of an expression.

ln :: ModelExpr -> ModelExpr

Smart constructor to take the ln of an expression.

sqrt :: ModelExpr -> ModelExpr

Smart constructor to take the square root of an expression.

sin :: ModelExpr -> ModelExpr

Smart constructor to apply sin to an expression.

cos :: ModelExpr -> ModelExpr

Smart constructor to apply cos to an expression.

tan :: ModelExpr -> ModelExpr

Smart constructor to apply tan to an expression.

sec :: ModelExpr -> ModelExpr

Smart constructor to apply sec to an expression.

csc :: ModelExpr -> ModelExpr

Smart constructor to apply csc to an expression.

cot :: ModelExpr -> ModelExpr

Smart constructor to apply cot to an expression.

arcsin :: ModelExpr -> ModelExpr

Smart constructor to apply arcsin to an expression.

arccos :: ModelExpr -> ModelExpr

Smart constructor to apply arccos to an expression.

arctan :: ModelExpr -> ModelExpr

Smart constructor to apply arctan to an expression.

exp :: ModelExpr -> ModelExpr

Smart constructor for the exponential (base e) function.

dim :: ModelExpr -> ModelExpr

Smart constructor for calculating the dimension of a vector.

norm :: ModelExpr -> ModelExpr

Smart constructor for calculating the normal form of a vector.

negVec :: ModelExpr -> ModelExpr

Smart constructor for negating vectors.

not_ :: ModelExpr -> ModelExpr

Smart constructor for applying logical negation to an expression.

idx :: ModelExpr -> ModelExpr -> ModelExpr

Smart constructor for indexing.

int :: Integer -> ModelExpr

Smart constructor for integers.

dbl :: Double -> ModelExpr

Smart constructor for doubles.

exactDbl :: Integer -> ModelExpr

Smart constructor for exact doubles.

space :: Space -> ModelExpr

Bring a Space into the ModelExpr.

frac :: Integer -> Integer -> ModelExpr

Smart constructor for fractions.

recip_ :: ModelExpr -> ModelExpr

Smart constructor for rational expressions (only in 1/x form).

str :: String -> ModelExpr

Smart constructor for strings.

perc :: Integer -> Integer -> ModelExpr

Smart constructors for percents.

defint :: Symbol -> ModelExpr -> ModelExpr -> ModelExpr -> ModelExpr

Smart constructor for the summation, product, and integral functions over an interval.

defsum :: Symbol -> ModelExpr -> ModelExpr -> ModelExpr -> ModelExpr

Smart constructor for the summation, product, and integral functions over an interval.

defprod :: Symbol -> ModelExpr -> ModelExpr -> ModelExpr -> ModelExpr

Smart constructor for the summation, product, and integral functions over an interval.

intAll :: Symbol -> ModelExpr -> ModelExpr

Smart constructor for the summation, product, and integral functions over all Real numbers.

sumAll :: Symbol -> ModelExpr -> ModelExpr

Smart constructor for the summation, product, and integral functions over all Real numbers.

prodAll :: Symbol -> ModelExpr -> ModelExpr

Smart constructor for the summation, product, and integral functions over all Real numbers.

realInterval :: HasUID c => c -> RealInterval ModelExpr ModelExpr -> ModelExpr

Smart constructor for 'real interval' membership.

euclidean :: [ModelExpr] -> ModelExpr

Euclidean function : takes a vector and returns the sqrt of the sum-of-squares.

sum' :: (Num a, Foldable t) => t a -> a

Used by euclidean function (in place of sum) to fix representation of computation.

cross :: ModelExpr -> ModelExpr -> ModelExpr

Smart constructor to cross product two expressions.

completeCase :: [(ModelExpr, ModelExpr)] -> ModelExpr

Smart constructor for case statement with complete set of cases.

incompleteCase :: [(ModelExpr, ModelExpr)] -> ModelExpr

Smart constructor for case statement with incomplete set of cases.

square :: ModelExpr -> ModelExpr

Smart constructor to square a function.

half :: ModelExpr -> ModelExpr

Smart constructor to half a function exactly.

oneHalf :: ModelExpr

Constructs 1/2.

oneThird :: ModelExpr

Constructs 1/3.

m2x2 :: ModelExpr -> ModelExpr -> ModelExpr -> ModelExpr -> ModelExpr

Matrix helper function.

vec2D :: ModelExpr -> ModelExpr -> ModelExpr

Matrix helper function.

dgnl2x2 :: ModelExpr -> ModelExpr -> ModelExpr

Matrix helper function.

apply :: (HasUID f, HasSymbol f) => f -> [ModelExpr] -> ModelExpr

Applies a given function with a list of parameters.

apply1 :: (HasUID f, HasSymbol f, HasUID a, HasSymbol a) => f -> a -> ModelExpr

Similar to apply, but converts second argument into Symbols.

apply2 :: (HasUID f, HasSymbol f, HasUID a, HasSymbol a, HasUID b, HasSymbol b) => f -> a -> b -> ModelExpr

Similar to apply, but the applied function takes two parameters (which are both Symbols).

applyWithNamedArgs :: (HasUID f, HasSymbol f, HasUID a, IsArgumentName a) => f -> [ModelExpr] -> [(a, ModelExpr)] -> ModelExpr

Similar to apply, but takes a relation to apply to FCall.

sy :: (HasUID c, HasSymbol c) => c -> ModelExpr

Get an ModelExpr from a Symbol.

deriv :: (HasUID c, HasSymbol c) => ModelExpr -> c -> ModelExpr

Gets the derivative of an ModelExpr with respect to a Symbol.

pderiv :: (HasUID c, HasSymbol c) => ModelExpr -> c -> ModelExpr

Gets the derivative of an ModelExpr with respect to a Symbol.

defines :: (Express a, Express b) => a -> b -> ModelExpr

One expression is "defined" by another.

isIn :: Express a => a -> ModelExpr -> ModelExpr

assocCreate :: Express d => AssocBoolOper -> [d] -> ModelExpr

Helper for creating new smart constructors for Associative Binary operations that require at least 1 expression.

assocSanitize :: AssocBoolOper -> [ModelExpr] -> [ModelExpr]

Helper for associative operations, removes embedded variants of the same kind

andMEs :: Express d => [d] -> ModelExpr

Binary associative And.

equivMEs :: Express a => [a] -> ModelExpr

Binary associative Equivalence.