Language.Drasil.NounPhrase

Types

class NounPhrase n where

Methods

phraseNP :: n -> Sentence

Retrieves singular form of term. Ex. "the quick brown fox".

pluralNP :: n -> PluralForm

Retrieves plural form of term. Ex. "the quick brown foxes".

sentenceCase :: n -> (NP -> Sentence) -> Capitalization

Retrieves the singular form and applies a captalization rule (usually capitalizes the first word) to produce a Sentence. Ex. "The quick brown fox".

titleCase :: n -> (NP -> Sentence) -> Capitalization

Retrieves the singular form and applies a captalization rule (usually capitalizes all words) to produce a Sentence. Ex. "The Quick Brown Fox".

data NP

For nouns and NounPhrases. May be constructed from a proper noun, common noun, or phrase (Sentence) and their respective pluralization and capitalization rules.

Phrase Accessors

atStartNP :: NounPhrase n => n -> Capitalization

Singular sentence case.

Helper function for getting the sentence case of a noun phrase.

atStartNP' :: NounPhrase n => n -> Capitalization

Plural sentence case.

Helper function for getting the sentence case of a noun phrase.

titleizeNP :: NounPhrase n => n -> Capitalization

Singular title case.

Helper function for getting the title case of a noun phrase.

titleizeNP' :: NounPhrase n => n -> Capitalization

Plural title case.

Helper function for getting the title case of a noun phrase.

Constructors

Common Noun Constructors

cn :: String -> NP

Self plural.

Constructs a common noun which capitalizes the first letter of the first word at the beginning of a sentence.

cn' :: String -> NP

Plural form simply adds "s" (ex. dog -> dogs).

Constructs a common noun which capitalizes the first letter of the first word at the beginning of a sentence.

cn'' :: String -> NP

Plural form adds "e" (ex. formula -> formulae).

Constructs a common noun which capitalizes the first letter of the first word at the beginning of a sentence.

cn''' :: String -> NP

Plural form adds "es" (ex. bush -> bushes).

Constructs a common noun which capitalizes the first letter of the first word at the beginning of a sentence.

cnICES :: String -> NP

Construct a common noun that pluralizes by dropping the last two letters and adding an "ices" ending (ex. matrix -> matrices).

cnIES :: String -> NP

Constructs a common noun that pluralizes by dropping the last letter and adding an "ies" ending (ex. body -> bodies).

cnIP :: String -> PluralRule -> NP

Constructs a common noun that allows you to specify the pluralization rule (as in pnIrr).

cnIS :: String -> NP

Constructs a common noun that pluralizes by dropping the last two letters and adding "es" (ex. analysis -> analyses).

cnIrr :: String -> PluralRule -> CapitalizationRule -> NP

Common noun that allows you to specify both the pluralization rule and the capitalization rule for sentence case (if the noun is used at the beginning of a sentence).

cnUM :: String -> NP

Constructs a common noun that pluralizes by dropping the last two letters and adding "a" (ex. datum -> data).

Proper Noun Constructors

pn :: String -> NP

Self plural.

Constructs a Proper Noun, it is always capitalized as written.

pn' :: String -> NP

Plural form simply adds "s" (ex. Henderson -> Hendersons).

Constructs a Proper Noun, it is always capitalized as written.

pn'' :: String -> NP

Plural form adds "e".

Constructs a Proper Noun, it is always capitalized as written.

pn''' :: String -> NP

Plural form adds "es" (ex. Bush -> Bushes).

Constructs a Proper Noun, it is always capitalized as written.

pnIrr :: String -> PluralRule -> NP

Constructs a ProperNoun with a custom plural rule (using IrregPlur from PluralRule). First argument is the String representing the noun, second is the rule.

Noun Phrase Constructors

nounPhrase :: String -> PluralString -> NP

Creates a NP with a given singular and plural form (as Strings) that capitalizes the first letter of the first word for sentence case.

nounPhrase' :: String -> PluralString -> CapitalizationRule -> NP

Similar to nounPhrase, but takes a specified capitalization rule for the sentence case.

nounPhrase'' :: Sentence -> PluralForm -> CapitalizationRule -> CapitalizationRule -> NP

Custom noun phrase constructor that takes a singular form (Sentence), plural form (Sentence), sentence case capitalization rule, and title case capitalization rule.

nounPhraseSP :: String -> NP

For things that should not be pluralized (or are self-plural). Works like nounPhrase, but with only the first argument.

nounPhraseSent :: Sentence -> NP

Similar to nounPhrase, except it only accepts one Sentence. Used for Requirements, Assumptions, LikelyChanges, etc. to allow for referencing. Plural case is just AddS.

Combinators

compoundPhrase :: (NounPhrase a, NounPhrase b) => a -> b -> NP

Combine two noun phrases. The singular form becomes phrase from t1 followed by phrase of t2. The plural becomes phrase of t1 followed by plural of t2. Uses standard CapFirst sentence case and CapWords title case. For example: compoundPhrase system constraint will have singular form "system constraint" and plural "system constraints".

compoundPhrase' :: NP -> NP -> NP

Similar to compoundPhrase, but the sentence case is the same as the title case (CapWords).

compoundPhrase'' :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Similar to compoundPhrase', but accepts two functions that will be used to construct the plural form. For example, compoundPhrase'' plural phrase system constraint would have the plural form "systems constraint".

compoundPhrase''' :: (NP -> Sentence) -> NP -> NP -> NP

Similar to compoundPhrase, but used when you need a special function applied to the first term of both singular and pluralcases (eg. short or plural).

compoundPhraseP1 :: NP -> NP -> NP

Similar to compoundPhrase, but pluralizes the first NP for both singular and plural cases.

Re-exported Types

data CapitalizationRule

Capitalization rules.

Constructors

CapFirst
Capitalize the first letter of the first word only.

CapWords
Capitalize the first letter of each word.

Replace Sentence
Replace the noun phrase with the given Sentence. Used for custom capitalization.

data PluralRule

Pluralization rules.

Constructors

AddS
Add "s" to the end of the noun phrase.

AddE
Add "e" to the end of the noun phrase.

AddES
Add "es" to the end of the noun phrase.

SelfPlur
The noun phrase is already plural.

IrregPlur (String -> String)
Apply the given function to the noun phrase to get the plural.