Language.Drasil.NounPhrase.Combinators

Description

Defines various NounPhrase-level combinators. These hold more information than those of the Sentence level but may not necessarily have a conceptual link. See the Wiki for more information and details about the naming process for combinators. A summary of the function naming scheme is as follows:

Combinators that conflict with haskell-native functions have an underscore appended.

Default plural case for combinators will be first term singular, second term plural.

P and S denote the plural case of the combinator when it does not follow the above default.

Gen denotes the general function case.

This module should be used as a qualified import (usually as NP), as many function names clash with those in Concepts.hs and Sentence.hs.

General Combinator Helper Functions

insertString :: String -> NP -> NP -> NP

Helper function that places a String in between two NPs. Plural case is (phraseNP t1) +:+ S s +:+ (pluralNP t2).

prependString :: String -> NP -> NP

Helper function that prepends a String to a NP.

insertSent :: Sentence -> NP -> NP -> NP

Helper function that places a Sentence in between two NPs. Plural case is (phraseNP t1) +:+ s +:+ (pluralNP t2).

prependSent :: Sentence -> NP -> NP

Helper function that prepends a Sentence to a NP.

Prepositions

"The" Combinators

the :: NP -> NP

Prepends "the" to a NP.

theGen :: (NP -> Sentence) -> NP -> NP

Similar to the, but accepts a function that determines the plural case.

"A" Combinators

a_ :: NP -> NP

Prepends "a" to a NP.

a_Gen :: (NP -> Sentence) -> NP -> NP

Similar to a, but accepts a function that determines the plural case.

Conjunctions

"And" Combinators

and_ :: NP -> NP -> NP

Inserts "and" between two NPs. Plural case is (phraseNP t1) +:+ "and" +:+ (pluralNP t2).

and_PS :: NP -> NP -> NP

Same as and_, but plural case is now (pluralNP t1) and_ (phraseNP t2).

and_Gen :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as and_, but takes two functions for the plural case.

and_GenGen :: (NP -> Sentence) -> (NP -> Sentence) -> (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as and_, but takes two functions for the singular case and two for the plural case.

andThe :: NP -> NP -> NP

Inserts "and the" between two NPs. Plural case is (phraseNP t1) +:+ "and the" +:+ (pluralNP t2).

"The" Combinators

ofThe :: NP -> NP -> NP

Inserts "of the" between two NPs. Plural case is (phraseNP t1) +:+ "of the" +:+ (pluralNP t2).

ofThePS :: NP -> NP -> NP

Similar to ofThe, but the plural case is now (pluralNP t1) ofThe (phraseNP t2).

ofTheGen :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Similar to ofThe, but accepts two functions for the plural case.

inThe :: NP -> NP -> NP

Inserts "in the" between two NPs. Plural case is (phraseNP t1) +:+ "in the" +:+ (pluralNP t2).

inThePS :: NP -> NP -> NP

Similar to ofThe, but the plural case is now (pluralNP t1) inThe (phraseNP t2).

inTheGen :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Similar to ofThe, but accepts two functions for the plural case.

the_ofThe :: NP -> NP -> NP

Prepends "the" and inserts "of the". Plural case is "the" +:+ (phraseNP t1) +:+ "of the" +:+ (pluralNP t2).

the_ofThePS :: NP -> NP -> NP

Similar to the_ofThe, but the plural case is now S "the" +:+ (pluralNP t1) ofThe (phraseNP t2).

the_ofTheGen :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Similar to the_ofThe', but takes two functions for the plural case.

"For" Combinators

for :: NP -> NP -> NP

Inserts "for" between two NPs. Plural case is (phraseNP t1) +:+ "for" +:+ (pluralNP t2).

forPS :: NP -> NP -> NP

Same as for, but plural case is now (pluralNP t1) for (phraseNP t2).

forGen :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as for', but takes two functions for the plural case.

"Of" Combinators

of_ :: NP -> NP -> NP

Inserts "of" between two NPs. Plural case is (phraseNP t1) +:+ "of" +:+ (pluralNP t2).

of_PS :: NP -> NP -> NP

Same as of_, but plural case is now (pluralNP t1) of_ (phraseNP t2).

of_Gen :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as of_, but takes two functions for the plural case.

of_GenGen :: (NP -> Sentence) -> (NP -> Sentence) -> (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as of_, but takes two functions for the singular case and two for the plural case.

Other Combinators

with :: NP -> NP -> NP

Inserts "with" between two NPs. Plural case is (phraseNP t1) +:+ "with" +:+ (pluralNP t2).