Language.Drasil.People

Description

Defines types and functions to encode people, names, and naming convention. Used for referencing and authorship of work.

Class

class HasName p

Members of this class must have a name.

Minimal complete definition

nameStr

Types

data Person

A person can have a given name, middle name(s), and surname, as well as the naming convention they use.

type People = [Person]

People is a synonymn for many Persons.

data Conv

Naming conventions.

Constructors

Western
Western style conventions are given name followed by middle names, followed by surname.

Eastern
Eastern style conventions are surname followed by middle names, followed by given name.

Mono
Mononyms are for those people who have only one name (ex. Madonna).

Constructors

person :: String -> String -> Person

Constructor for a person using Western naming conventions. Used for a person with only a given name and surname. Arguments are in the order: given name, surname.

person' :: String -> String -> Person

Constructor for a person using Eastern naming conventions. Used for a person with only a given name and surname. Arguments are in the order: surname, given name.

personWM :: String -> [String] -> String -> Person

Constructor for a person using Western naming conventions. Similar to the person constructor, except the middle argument is a list of middle names.

personWM' :: String -> [String] -> String -> Person

Constructor for a person using Eastern naming conventions. Similar to the person' constructor, except the middle argument is a list of middle names.

mononym :: String -> Person

Constructor for a person with a mononym (only one name).

Accessors

name :: HasName n => n -> String

Gets the name of a Person. Adds a dot after any initials.

nameStr :: HasName p => p -> String

Provides the ability to hold a name.

lstName :: Person -> String

Gets the last name of a Person.

rendPersLFM :: Person -> String

Gets a Person's name in the form: Last, First Middle.

rendPersLFM' :: Person -> String

Gets a Person's name in the form: Last, F. M.

rendPersLFM'' :: Person -> String

Gets a Person's name in the form: Last, First M.

comparePeople :: [Person] -> [Person] -> Ordering

Orderes different groups of Persons. If two lists are the same up to a point, the citation with more Persons will go last.