Language.Drasil.RefProg

Documentation

data Reference

A Reference contains the identifier (UID), a reference address (LblType), a human-readable shortname (ShortName), and any extra information about the reference (RefInfo).

Constructors

Reference UID LblType ShortName RefInfo

data RefInfo

Holds any extra information needed for a Reference, be it an equation, pages, a note, or nothing.

Constructors

None

Equation [Int]

Page [Int]

RefNote String