Language.Drasil.Sentence

Description

Contains Sentences and helpers

Documentation

data Sentence where

Constructors

Ch :: SentenceStyle -> UID -> Sentence

Sy :: USymb -> Sentence

S

Fields

:: String

-> Sentence

Strings, used for Descriptions in Chunks

P :: Symbol -> Sentence

E :: Expr -> Sentence

Ref :: Reference -> Sentence

Quote

Fields

:: Sentence

-> Sentence

Adds quotation marks around a sentence

Percent

Fields

:: Sentence

% symbol

(:+:) :: Sentence -> Sentence -> Sentence infixr 5
Direct concatenation of two Sentences (no implicit spaces!)

EmptyS :: Sentence

data SentenceStyle

For writing "sentences" via combining smaller elements. Sentences are made up of some known vocabulary of things:

units (their visual representation)

words (via String)

special characters

accented letters

References to specific layout objects

Constructors

PluralTerm

SymbolStyle

TermStyle

ShortStyle

(+:+) :: Sentence -> Sentence -> Sentence

Helper for concatenating two sentences with a space between them.

(+:+.) :: Sentence -> Sentence -> Sentence

Helper which concatenates two sentences using +:+ then adds a period to the end.

(+:) :: Sentence -> Sentence -> Sentence

Helper which concatenates two sentences using +:+ then adds a colon to the end.

(!.) :: Sentence -> Sentence

Helper which appends a period to the end of a sentence

capSent :: Sentence -> Sentence

Capitalizes a Sentence.

ch :: (HasUID c, HasSymbol c) => c -> Sentence

Gets a symbol and places in a Sentence

sC :: Sentence -> Sentence -> Sentence

Helper for concatenating two sentences with a comma and space between them.

sDash :: Sentence -> Sentence -> Sentence

Helper for concatenating two sentences with a space-surrounded dash between them.

sentencePlural :: UID -> Sentence

Smart constructors for turning a UID into a Sentence

sentenceShort :: UID -> Sentence

Smart constructors for turning a UID into a Sentence

sentenceSymb :: UID -> Sentence

Smart constructors for turning a UID into a Sentence

sentenceTerm :: UID -> Sentence

Smart constructors for turning a UID into a Sentence

sParen :: Sentence -> Sentence

Helper for wrapping sentences in parentheses.