Language.Drasil.Sentence.Combinators

Description

Defines various sentence level combinators that may be based in English and not necessarily have a conceptual link. See the Wiki for more information. There are also C variants to each combinator to denote those meant for use at the start of a sentence (capitalizes the first word). This module should be used as a qualified import (usually as S), as many function names clash with those in Concepts.hs and NounPhrase.hs.

"And" Combinators

and_ :: Sentence -> Sentence -> Sentence

Inserts the word "and" between two Sentences.

andIts :: Sentence -> Sentence -> Sentence

Inserts the words "and its" between two Sentences.

andThe :: Sentence -> Sentence -> Sentence

Inserts the words "and the" between two Sentences.

"The" Combinators

fromThe :: Sentence -> Sentence -> Sentence

Inserts the words "from the" between two Sentences.

inThe :: Sentence -> Sentence -> Sentence

Inserts the words "in the" between two Sentences.

onThe :: Sentence -> Sentence -> Sentence

Inserts the words "on the" between two Sentences.

toThe :: Sentence -> Sentence -> Sentence

Inserts the words "to the" between two Sentences.

isThe :: Sentence -> Sentence -> Sentence

Inserts the words "is the" between two Sentences.

ofThe :: Sentence -> Sentence -> Sentence

Inserts the words "of the" between two Sentences.

the_ofThe :: Sentence -> Sentence -> Sentence

Same as ofThe, but inserts "the" at the beginning of the Sentence.

the_ofTheC :: Sentence -> Sentence -> Sentence

Same as the_ofThe, except first "the" is capitalized.

the_ofGiv :: Sentence -> Sentence -> Sentence

Prepends "the" and inserts "of a given" between two Sentences.

the_ofGivC :: Sentence -> Sentence -> Sentence

Same as ofGiv, except first "the" is capitalized.

the_isExpctdToHvC :: Sentence -> Sentence -> Sentence

Prepends "The" and inserts "is expected to have" between two Sentences.

"For" Combinators

forTPS :: (NamedIdea c, NamedIdea d) => c -> d -> Sentence

Similar to for, but used for titles and first NamedIdea is pluralized.

forTPP :: (NamedIdea c, NamedIdea d) => c -> d -> Sentence

Similar to forTTPS, but both NamedIdeas are pluralized.

for :: Sentence -> Sentence -> Sentence

Inserts the word "for" between two Sentences.

forT :: (NamedIdea c, NamedIdea d) => c -> d -> Sentence

Similar to for, but both terms are titleized.

forGen :: (c -> Sentence) -> (d -> Sentence) -> c -> d -> Sentence

Similar to forTT, but takes two arguments (for capitalization or pluralization) to apply to the two terms respectively.

Other Combinators

of_ :: Sentence -> Sentence -> Sentence

Inserts the word "of" between two Sentences.

ofA :: Sentence -> Sentence -> Sentence

Inserts the words "of a" between two Sentences.

or_ :: Sentence -> Sentence -> Sentence

Inserts the word "or" between two Sentences.

are :: Sentence -> Sentence -> Sentence

Inserts the word "are" between two Sentences.

in_ :: Sentence -> Sentence -> Sentence

Inserts the word "in" between two Sentences.

is :: Sentence -> Sentence -> Sentence

Inserts the word "is" between two Sentences.

defnAs :: Sentence -> Sentence -> Sentence

Inserts the words "defined as" between two Sentences.

denotes :: Sentence -> Sentence -> Sentence

Inserts the words "denotes the" between two Sentences.

versus :: Sentence -> Sentence -> Sentence

Inserts the word "versus" between two Sentences.

wrt :: Sentence -> Sentence -> Sentence

Inserts the words "with respect to" between two Sentences.