Language.Drasil.Sentence.Extract

Description

Extract various kinds of UIDs from a Sentence. Used in conjunction with the chunk database in order to render terms, symbols, and references properly.

Documentation

sdep :: Sentence -> [UID]

This is to collect symbolic UIDs that are printed out as a Symbol.

shortdep :: Sentence -> [UID]

lnames :: Sentence -> [UID]

Generic traverse of all positions that could lead to reference UIDs from Sentences.

lnames' :: [Sentence] -> [UID]

Get reference UIDs from Sentences.