Language.Drasil.Sentence.Generators

Description

Functions that create sentences from other data

Documentation

fromSource :: (Referable r, HasShortName r) => r -> Sentence

Wraps "from reference" in parentheses.

fterms :: (NamedIdea c, NamedIdea d) => (NP -> NP -> t) -> c -> d -> t

Apply a binary function to the terms of two named ideas, instead of to the named ideas themselves. Ex. fterms compoundPhrase t1 t2 instead of compoundPhrase (t1 ^. term) (t2 ^. term).

getTandS :: Quantity a => a -> Sentence

Used when you want to say a term followed by its symbol. ex. "...using the Force F in...".

checkValidStr :: String -> String -> Either String String

Uses an Either type to check if a String is valid - Left with error message if there is an invalid Char in String, else Right with String.