Language.Drasil.ShortHands

Description

Alphabet of capital/lowercase English letters as symbols.

Documentation

cA :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cB :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cC :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cD :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cE :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cF :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cG :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cH :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cI :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cJ :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cK :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cL :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cM :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cN :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cO :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cP :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cQ :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cR :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cS :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cT :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cU :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cV :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cW :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cX :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cY :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cZ :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lA :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lB :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lC :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lD :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lE :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lF :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lG :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lH :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lI :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lJ :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lK :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lL :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lM :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lN :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lO :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lP :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lQ :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lR :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lS :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lT :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lU :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lV :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lW :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lX :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lY :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lZ :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lAlpha :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cAlpha :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lBeta :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cBeta :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lGamma :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cGamma :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lDelta :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cDelta :: Symbol

Δ is often used as a decoration on top of a symbol. For example, Δx and Δy represent the change in x and the change in y. The variable defined here represents delta as its own independent symbol and should not be used to represent "the change in" something.

Short forms of various variable names. c means capital; l means lowercase.

lEpsilon :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

vEpsilon :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cEpsilon :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lZeta :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cZeta :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lEta :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cEta :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lTheta :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cTheta :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lIota :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cIota :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lKappa :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cKappa :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lLambda :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cLambda :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lMu :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cMu :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lNu :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cNu :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lXi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cXi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lOmicron :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cOmicron :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lPi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cPi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lRho :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cRho :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lSigma :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cSigma :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lTau :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cTau :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lUpsilon :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cUpsilon :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lPhi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

vPhi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cPhi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lChi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cChi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lPsi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cPsi :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lOmega :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

cOmega :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lNabla :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lEll :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lPosInf :: Symbol

Short forms of various variable names. c means capital; l means lowercase.

lNegInf :: Symbol

Short forms of various variable names. c means capital; l means lowercase.