Language.Drasil.ShortName

Description

Short names are used for displaying references.

Type

newtype ShortName

Used for holding the short form of a name (as a Sentence with a wrapper).

Constructors

ShortNm Sentence

Class

class HasShortName s where

A ShortName is the text to be displayed for a link.

Used for referencing within a document that can include symbols and whatnot if required. Visible in the typeset documents (pdf).

Methods

shortname :: s -> ShortName

Functions

getSentSN :: ShortName -> Sentence

Pulls the short form (as a Sentence) out of a ShortName.

shortname' :: Sentence -> ShortName

Smart constructor for making a Sentence into a ShortName.