Language.Drasil.Space

Description

Number space types and functions.

Types

data Space

The difference kinds of spaces that may exist. This type holds numerical spaces (such as the set of integers, rationals, etc.), a space for booleans, a space for characters, dimensional spaces (vectors, arrays, etc.), a space for Actors, discrete sets (both for numbers and strings), and a void space.

Constructors

Integer

Rational

Real

Natural

Boolean

Char

String

Radians

Vect Space

Array Space

Actor String

DiscreteD [Double]

DiscreteS [String]

Function (NonEmpty Primitive) Primitive

Void

data RealInterval a b where

A RealInterval is a subset of Real (as a Space). These come in different flavours. For now, we embed Expr for the bounds, but that will change as well.

Constructors

Bounded

Fields

:: (Inclusive, a)

-> (Inclusive, b)

-> RealInterval a b

Interval from (x .. y).

UpTo

Fields

:: (Inclusive, a)

-> RealInterval a b

Interval from (-infinity .. x).

UpFrom

Fields

:: (Inclusive, b)

-> RealInterval a b

Interval from (x .. infinity).

data Inclusive

Inclusive or exclusive bounds.

Constructors

Inc

Exc

data DomainDesc (tplgy :: RTopology) a b where

Describes the domain of a Symbol given a topology. Can be bounded or encase all of the domain.

Constructors

BoundedDD :: Symbol -> RTopology -> a -> b -> DomainDesc 'Discrete a b

AllDD :: Symbol -> RTopology -> DomainDesc 'Continuous a b

data RTopology

Topology of a subset of reals.

Constructors

Continuous

Discrete

type DiscreteDomainDesc a b = DomainDesc 'Discrete a b

type ContinuousDomainDesc a b = DomainDesc 'Continuous a b

Class

class HasSpace c where

HasSpace is anything which has a Space.

Methods

typ :: Lens' c Space

Provides a Lens to the Space.

Functions

getActorName :: Space -> String

Gets the name of an Actor.

getInnerSpace :: Space -> Space

Gets the inner Space of a vector.

mkFunction :: [Primitive] -> Primitive -> Space