Language.Drasil.Symbol.Helpers

Description

Routines to help with Symbols and Stages.

Documentation

eqSymb :: HasSymbol q => q -> Symbol

Helper function for getting a symbol in the Equational Stage.

codeSymb :: HasSymbol q => q -> Symbol

Helper function for getting a symbol in the Implementation Stage.

hasStageSymbol :: HasSymbol q => q -> Stage -> Bool

Finds if a Stage symbol is real or Empty. True if real.

autoStage :: Symbol -> Stage -> Symbol

Helper for creating a symbol with Unicode in it.

hat :: Symbol -> Symbol

Helper for creating a symbol with a hat ("^") atop it.

prime :: Symbol -> Symbol

Helper for creating a Vector symbol.

staged :: Symbol -> Symbol -> Stage -> Symbol

Helper for creating a symbol that depends on the stage.

sub :: Symbol -> Symbol -> Symbol

Helper for creating a symbol with a subscript to the right. Arguments: Base symbol, then subscripted symbol.

subStr :: Symbol -> String -> Symbol

Helper for a common case of subscript, with a string Arguments: Base symbol, then subscript String.

sup :: Symbol -> Symbol -> Symbol

Helper for creating a symbol with a superscript to the right. Arguments: Base symbol, then superscripted symbol.

unicodeConv :: Symbol -> Symbol

Helper for autoStage that applies unicodeString to all Symbols with Strings.

upperLeft :: Symbol -> Symbol -> Symbol

Helper for creating a symbol with a superscript on the left side of the symbol. Arguments: Base symbol, then superscripted symbol.

vec :: Symbol -> Symbol

Helper for creating a Vector symbol.

label :: String -> Symbol

Label smart constructor, requires non-empty labels

variable :: String -> Symbol

Variable smart constructor, requires non-empty variables