Language.Drasil.UID

Description

Unique Identifier used across Drasil.

Documentation

data UID

A UID is a 'unique identifier' for things that we will put into our database of information. We use a newtype wrapper to make sure we are only using UIDs where desired.

class HasUID c where

The most basic item: having a unique identifier key, here a UID.

Methods

uid :: Lens' c UID

Provides a unique id for internal Drasil use.

mkUid :: String -> UID

Smart constructor for making a UID from a String.

(+++) :: HasUID a => a -> String -> UID

For when we need to modify a UID. We first take the base chunk's UID and then append a suffix to it.

(+++.) :: UID -> String -> UID

For when we need to append something to a UID.

(+++!) :: (HasUID a, HasUID b) => a -> b -> UID

showUID :: HasUID a => a -> String

Grabs the UID from something that has a UID and displays it as a String.