Language.Drasil.Uncertainty

Description

Uncertainty functions.

Constructor

defaultUncrt :: Uncertainty

The default uncertainty is set to 0.1.

Accessors

uncVal :: HasUncertainty x => x -> Double

Accessor for uncertainty value from something that has an uncertainty.

uncPrec :: HasUncertainty x => x -> Maybe Int

Accessor for precision value from something that has an uncertainty.

Re-exported

uncty :: Double -> Maybe Int -> Uncertainty

Smart constructor for values with uncertainty.

exact :: Uncertainty

Smart constructor for exact values (no uncertainty).