Language.Drasil.UnitLang

Description

Language for defining and manipulating units.

Types

newtype USymb

Language of units (how to build them up into a unit symbol). Of the form (Symbol ^ Integer). The Integer may be negative, but should not be zero.

Constructors

US [(Symbol, Integer)]

data UDefn

Language of unit equations, to define a unit relative to another.

Constructors

USynonym USymb
to define straight synonyms.

UScale Double USymb
scale, i.e. *.

UShift Double USymb
shift, i.e. +.

data UnitSymbol

When we define units, they come in three flavours: SI (base) units, derived SI units (aka synonyms), and defined units. The type below captures that knowledge.

Constructors

BaseSI USymb

DerivedSI USymb USymb UDefn

Defined USymb UDefn

Functions

fromUDefn :: UDefn -> USymb

Generates a default unit symbol.

compUSymb :: USymb -> USymb -> Ordering

Hand-rolled version of compare. Should assume USymb is normalized, so that some redundant EQ cases can be removed.

getUSymb :: UnitSymbol -> USymb

Generates a default unit symbol.

getDefn :: UnitSymbol -> Maybe UDefn

Gets the unit definition of a unit symbol.