Language.Drasil.WellTyped

Documentation

type TypeError = String

type TypingContext t = Map UID t

We can only type check UIDs within a type context relating UIDs to types since they don't carry any type information.

inferFromContext :: TypingContext t -> UID -> Either TypeError t

Look for a known type of a specific UID.

class (Eq t, Show t) => Typed e t where

Build a bidirectional type checker for your expression language, e, with respect to a specific type universe, t.

Methods

infer :: TypingContext t -> e -> Either TypeError t

Given a typing context and an expression, infer a unique type or explain what went awry.

check :: TypingContext t -> e -> t -> Either TypeError t

Given a typing context, an expression, and an expected type, check if the expression can satisfy the expectation.

class Typed e t => RequiresChecking c e t where

For all containers, c, which contain typed expressions, e, against a specific type universe, t, expose all expressions and relations that need to be type-checked.

Methods

requiredChecks :: c -> [(e, t)]

All things that need type checking.

assertEq :: (Show t, Eq t) => t -> t -> (String -> String -> TypeError) -> Either TypeError ()

(~==) :: (Show t, Eq t) => t -> t -> (String -> String -> TypeError) -> Either TypeError () infix 4

typeCheckByInfer :: Typed e t => TypingContext t -> e -> t -> Either TypeError t

`Check' an expressions type based by an inference.

assertAllEq :: Typed e t => TypingContext t -> [e] -> t -> TypeError -> Either TypeError ()

temporaryIndent :: String -> String -> String

A temporary, hacky, indentation function. It should be removed when we switch to using something else for error messages, which can be later formatted nicely.