Language.Drasil.Chunk.Concept

Description

Contains functions to create the concept related chunk types found in Language.Drasil.Chunk.Concept.Core.

Concept Chunks

From an idea (IdeaDict)

data ConceptChunk

The ConceptChunk datatype records a concept that contains an idea (IdeaDict), a definition (Sentence), and an associated domain of knowledge ([UID]).

Ex. The concept of Accuracy may be defined as the quality or state of being correct or precise.

dcc :: String -> NP -> String -> ConceptChunk

Smart constructor for creating concept chunks given a UID, NounPhrase (NP) and definition (as a String).

Concept domain tagging is not yet implemented in this constructor.

dccA :: String -> NP -> String -> Maybe String -> ConceptChunk

Smart constructor for creating a concept chunks with an abbreviation Takes a UID (String), a term (NounPhrase), a definition (String), and an abbreviation (Maybe String).

dccAWDS :: String -> NP -> Sentence -> Maybe String -> ConceptChunk

dccWDS :: String -> NP -> Sentence -> ConceptChunk

Similar to dcc, except the definition takes a Sentence.

cc :: Idea c => c -> String -> ConceptChunk

Constructor for projecting an idea into a ConceptChunk. Takes the definition of the ConceptChunk as a String. Does not allow concept domain tagging.

cc' :: Idea c => c -> Sentence -> ConceptChunk

Same as cc, except definition is a Sentence.

ccs :: (Idea c, Concept d) => c -> Sentence -> [d] -> ConceptChunk

Similar to cc', but allows explicit domain tagging.

cw :: Concept c => c -> ConceptChunk

For projecting out to the ConceptChunk data-type.

From a ConceptChunk

data ConceptInstance

Contains a ConceptChunk, reference address, and a ShortName. It is a concept that can be referred to, or rather, a instance of where a concept is applied. Often used in Goal Statements, Assumptions, Requirements, etc.

Ex. Something like the assumption that gravity is 9.81 m/s. When we write our equations, we can then link this assumption so that we do not have to explicitly define that assumption when needed to verify our work.

cic :: Concept c => String -> Sentence -> String -> c -> ConceptInstance

Constructor for a ConceptInstance. Takes in the Reference Address (String), a definition (Sentence), a short name (String), and a domain (for explicit tagging).