Language.Drasil.Chunk.Concept.Core

Description

Define concept-related chunks. A concept is usually something that has a term, definition, and comes from some domain of knowledge.

Concept-related Datatypes

data ConceptChunk

The ConceptChunk datatype records a concept that contains an idea (IdeaDict), a definition (Sentence), and an associated domain of knowledge ([UID]).

Ex. The concept of Accuracy may be defined as the quality or state of being correct or precise.

Constructors

ConDict IdeaDict Sentence [UID]

data ConceptInstance

Contains a ConceptChunk, reference address, and a ShortName. It is a concept that can be referred to, or rather, a instance of where a concept is applied. Often used in Goal Statements, Assumptions, Requirements, etc.

Ex. Something like the assumption that gravity is 9.81 m/s. When we write our equations, we can then link this assumption so that we do not have to explicitly define that assumption when needed to verify our work.

Constructors

ConInst UID ConceptChunk String ShortName

sDom :: [UID] -> UID

Check if something has one domain. Throws an error if there is more than one.