Language.Drasil.Chunk.Eq

Description

Contains chunks related to adding an expression to a quantitative concept.

Chunk Type

data QDefinition

A QDefinition is a QuantityDict with a defining expression (Expr), a definition (Sentence), and a domain ([UID]). This chunk now contains almost enough information to generate code, definitions, and models. Used for high-level quantities with equations and inputs.

Constructor

fromEqn :: IsUnit u => UID -> NP -> Sentence -> Symbol -> Space -> u -> Expr -> QDefinition

Create a QDefinition with a UID, term (NP), definition (Sentence), Symbol, Space, unit, and defining expression.

fromEqn' :: String -> NP -> Sentence -> Symbol -> Space -> Expr -> QDefinition

Same as fromEqn, but has no units.

fromEqnSt :: IsUnit u => String -> NP -> Sentence -> (Stage -> Symbol) -> Space -> u -> Expr -> QDefinition

Same as fromEqn, but symbol depends on stage.

fromEqnSt' :: String -> NP -> Sentence -> (Stage -> Symbol) -> Space -> Expr -> QDefinition

Same as fromEqn, but symbol depends on stage and has no units.

mkQDefSt :: UID -> NP -> Sentence -> (Stage -> Symbol) -> Space -> Maybe UnitDefn -> Expr -> QDefinition

Wrapper for fromEqnSt and fromEqnSt'

mkQuantDef :: (Quantity c, MayHaveUnit c) => c -> Expr -> QDefinition

Used to help make QDefinitions when UID, term, and Symbol come from the same source.

mkQuantDef' :: (Quantity c, MayHaveUnit c) => c -> NP -> Expr -> QDefinition

Used to help make QDefinitions when UID and Symbol come from the same source, with the term separate.

ec :: (Quantity c, MayHaveUnit c) => c -> Expr -> QDefinition

Smart constructor for QDefinitions. Requires a quantity and its defining equation.

mkFuncDef :: (HasUID f, HasSymbol f, HasSpace f, HasUID i, HasSymbol i, HasSpace i, IsUnit u) => f -> NP -> Sentence -> u -> [i] -> Expr -> QDefinition

Create a QDefinition function with a symbol, name, term, list of inputs, resultant units, and a defining Expr

mkFuncDef' :: (HasUID f, HasSymbol f, HasSpace f, HasUID i, HasSymbol i, HasSpace i) => f -> NP -> Sentence -> [i] -> Expr -> QDefinition

Create a QDefinition function with a symbol, name, term, list of inputs, and a defining Expr

mkFuncDefByQ :: (Quantity c, MayHaveUnit c, HasSpace c, Quantity i, HasSpace i) => c -> [i] -> Expr -> QDefinition

Create a QDefinition functions using a symbol, list of inputs, and a defining Expr