Language.Drasil.Chunk.NamedIdea

Description

The lowest level of chunks in Drasil. It all starts with an identifier and a term.

Chunk Types

data NamedChunk

Used for anything worth naming. Note that a NamedChunk does not have an acronym/abbreviation as that's a CommonIdea, which has its own representation. Contains a UID and a term that we can capitalize or pluralize (NP).

Ex. Anything worth naming must start out somewhere. Before we can assign equations and values and symbols to something like the arm of a pendulum, we must first give it a name.

data IdeaDict

IdeaDict is the canonical dictionary associated to an Idea. Contains a NamedChunk that could have an abbreviation (Maybe String).

Ex. The project name "Double Pendulum" may have the abbreviation DblPendulum.

Constructors

nc :: UID -> NP -> NamedChunk

NamedChunk constructor, takes a UID and a term.

nw :: Idea c => c -> IdeaDict

Historical name: nw comes from 'named wrapped' from when NamedIdea exported getA (now in Idea). But there are no more wrappers, instead we have explicit dictionaries. Unwraps an Idea and places its UID and NP into an IdeaDict with Nothing for an abbreviation.

mkIdea :: UID -> NP -> Maybe String -> IdeaDict

IdeaDict constructor, takes a UID, NP, and an abbreviation in the form of Maybe String