Language.Drasil.Chunk.Relation

Description

For adding a relation (expression) to a concept.

Chunk Type

data RelationConcept

For a concept (ConceptChunk) that also has a Relation (ModelExpr) attached.

Ex. We can describe a pendulum arm and then apply an associated equation so that we know its behaviour.

Constructors

makeRC :: Express e => String -> NP -> Sentence -> e -> RelationConcept

Create a RelationConcept from a given UID, term (NP), definition (Sentence), and Relation.