Language.Drasil.Chunk.UncertainQuantity

Description

For adding an uncertainty value to quantities with constraints.

Chunk Types

data UncertQ

UncertQs are conceptual symbolic quantities with constraints and an Uncertainty. Contains a ConstrConcept and an Uncertainty.

Ex. Measuring the length of a pendulum arm may be recorded with an uncertainty value.

Constructors

uq :: (Quantity c, Constrained c, Concept c, HasReasVal c, MayHaveUnit c) => c -> Uncertainty -> UncertQ

Smart constructor that requires a Quantity, a percentage, and a typical value with an Uncertainty.

uqc :: IsUnit u => String -> NP -> String -> Symbol -> u -> Space -> [ConstraintE] -> Expr -> Uncertainty -> UncertQ

Uncertainty quantity (uq) but with a constraint.

uqcND :: IsUnit u => String -> NP -> Symbol -> u -> Space -> [ConstraintE] -> Expr -> Uncertainty -> UncertQ

Uncertainty quantity constraint (uqc) without a description.