Language.Drasil.Chunk.Unitary

Description

Defines chunks to add units to a quantity. Similar to UnitalChunk.

Chunk Types

class Quantity c => Unitary c where

A Unitary is a Quantity that must have a unit.

Methods

unit :: c -> UnitDefn

Helpers

unit_symb :: Unitary c => c -> USymb

Helper for getting the unit's Symbol from a chunk, as opposed to the symbols of the chunk itself.