Language.Drasil.Classes

Description

Defining all the classes which represent knowledge-about-knowledge.

Classes

Chunks

class HasUID c => NamedIdea c where

A NamedIdea is a term that we've identified (has a UID) as being worthy of naming.

Methods

term :: Lens' c NP

Lens to the term (a noun phrase).

class NamedIdea c => Idea c where

An Idea is the combination of a NamedIdea and a CommonIdea. In other words, it may have an acronym/abbreviation.

Methods

getA :: c -> Maybe String

Gets the acronym/abbreviation.

class NamedIdea c => CommonIdea c where

CommonIdea is a NamedIdea with the additional constraint that it must have an abbreviation. This is the main distinction between getA and abrv, where getA may return Nothing, while abrv will always return the abbreviation.

Methods

abrv :: c -> String

Introduces abrv which necessarily provides an abbreviation.

type Concept c = (Idea c, Definition c, ConceptDomain c)

Concepts are Ideas with definitions and domains.

class Definition c where

Defines a chunk.

Methods

defn :: Lens' c Sentence

Provides (a Lens to) the definition for a chunk.

class ConceptDomain c where

Some concepts have a domain (related information encoded in UIDs to other chunks).

Methods

cdom :: c -> [UID]

Provides Getter for the concept domain tags for a chunk

cdom should be exported for use by the Drasil framework, but should not be exported beyond that.

class (Idea c, HasSpace c, HasSymbol c) => Quantity c

A Quantity is an Idea with a Space and a Symbol. In theory, it should also restrict to being a part of MayHaveUnit, but that causes all sorts of import cycles (or lots of orphans).

class HasUnitSymbol u where

Some chunks store a unit symbol.

Methods

usymb :: u -> USymb

Provides the ability to hold a unit symbol (USymb).

class HasReasVal c where

A Quantity that could have a reasonable value.

Methods

reasVal :: Lens' c (Maybe Expr)

Provides a Lens to the possible reasonable value.

class Constrained c where

The Constrained class is a Quantity that has value constraints. It does not enforce Quantity at this point.

Methods

constraints :: Lens' c [ConstraintE]

Provides a Lens to the Constraints.

class HasAdditionalNotes c where

Records any additional notes needed to avoid losing information

Methods

getNotes :: Lens' c [Sentence]

Provides a Lens to the notes.

class (Idea u, Definition u, HasUnitSymbol u) => IsUnit u where

Units are Ideas with a Definition which store a unit symbol. They must also be explicitly declared to be instances of IsUnit.

Methods

udefn :: u -> Maybe UDefn

May have a unit definition.

getUnits :: u -> [UID]

Holds units as a list of UID.

class UnitEq u where

Contains a UDefn

Methods

uniteq :: Lens' u UDefn

Provides the Lens to a unit definition.

Expr and expressions

class Express c where

Data that can be expressed using ModelExpr.

Methods

express :: c -> ModelExpr

class DefiningExpr c where

Methods

defnExpr :: Lens' (c e) e

Provides a Lens to the expression. TODO: Well, technically, e doesn't need to be an "expression" of any sorts. It just needs to be _something_, and it would have approximately have same meaning.