Language.Drasil.Classes.Core

Description

Defining the core classes which represent knowledge-about-knowledge.

Documentation

class HasUID c where

The most basic item: having a unique identifier key, here a UID.

Methods

uid :: Lens' c UID

Provides a unique id for internal Drasil use.

class HasRefAddress b where

Members must have a reference address.

Methods

getRefAdd :: b -> LblType

Provides the ability to hold a reference address.

class HasSymbol c where

A HasSymbol is anything which has a Symbol.

Methods

symbol :: c -> Stage -> Symbol

Provides the Symbol for a particular stage of generation.

class (HasUID s, HasRefAddress s) => Referable s where

Members of this class have the ability to be referenced.

Methods

refAdd :: s -> String

The referencing address (what we're linking to). Only visible in the source (tex/html).

renderRef :: s -> LblType

Alternate form of reference.