Language.Drasil.Development

Documentation

class NounPhrase n where

Minimal complete definition

phraseNP, pluralNP, sentenceCase, titleCase

Methods

phraseNP

Arguments

:: n

-> Sentence
ex. "the quick brown fox"

pluralNP

Arguments

:: n

-> PluralForm
ex. "the quick brown foxes" Could replace plural string with a function.

data UFuncB

Constructors

Not

data UFuncVec

Constructors

Norm

Dim

data ArithBinOp

Constructors

Frac

Pow

Subt

data BoolBinOp

Constructors

Impl

Iff

data EqBinOp

Constructors

Eq

NEq

data LABinOp

Constructors

Index

data OrdBinOp

Constructors

Lt

Gt

LEq

GEq

data VVVBinOp

Constructors

Cross

data VVNBinOp

Constructors

Dot

dep :: Expr -> [String]

Get dependencies from an equation.

names :: Expr -> [String]

Generic traverse of all expressions that could lead to names.

names' :: Expr -> [String]

Generic traverse of all positions that could lead to names without functions. FIXME : this should really be done via post-facto filtering, but right now the information needed to do this is not available!

namesRI :: RealInterval Expr Expr -> [String]

Generic traversal of everything that could come from an interval to names. (Similar to names)

sdep :: Sentence -> [UID]

This is to collect UIDs that are printed out as a Symbol.

lnames :: Sentence -> [UID]

Generic traverse of all positions that could lead to reference UIDs from Sentences.

lnames' :: [Sentence] -> [UID]

Get reference UIDs from Sentences.

precA :: AssocArithOper -> Int

precA - precedence for arithmetic-related Binary-Associative (Commutative) operators.

precB :: AssocBoolOper -> Int

precB - precedence for boolean-related Binary-Associative (Commutative) operators.

eprec :: Expr -> Int

eprec - Expression precedence.