Language.Drasil.Development

Description

Developing the expression language in Drasil. Re-export many things to simplify external use.

NounPhrase

class NounPhrase n where

Minimal complete definition

phraseNP, pluralNP, sentenceCase, titleCase

Methods

phraseNP :: n -> Sentence

Retrieves singular form of term. Ex. "the quick brown fox".

pluralNP :: n -> PluralForm

Retrieves plural form of term. Ex. "the quick brown foxes".

Expr

data Expr where

Drasil expressions.

Constructors

Dbl :: Double -> Expr
Turns a decimal value (Double) into an expression.

Int :: Integer -> Expr
Turns an integer into an expression.

ExactDbl :: Integer -> Expr
Represents decimal values that are exact as integers.

Str :: String -> Expr
Turns a string into an expression.

Perc :: Integer -> Integer -> Expr
Turns two integers into a fraction (or percent).

AssocA :: AssocArithOper -> [Expr] -> Expr
Takes an associative arithmetic operator with a list of expressions.

AssocB :: AssocBoolOper -> [Expr] -> Expr
Takes an associative boolean operator with a list of expressions.

Deriv :: DerivType -> Expr -> UID -> Expr
Derivative syntax is: Type (Partial or Total) -> principal part of change -> with respect to For example: Deriv Part y x1 would be (dy/dx1).

C :: UID -> Expr
C stands for Chunk, for referring to a chunk in an expression. Implicitly assumes that the chunk has a symbol.

FCall :: UID -> [Expr] -> [(UID, Expr)] -> Expr
A function call accepts a list of parameters and a list of named parameters. For example

F(x) is (FCall F [x] []).

F(x,y) would be (FCall F [x,y]).

F(x,n=y) would be (FCall F [x] [(n,y)]).

Case :: Completeness -> [(Expr, Relation)] -> Expr
For multi-case expressions, each pair represents one case.

Matrix :: [[Expr]] -> Expr
Represents a matrix of expressions.

UnaryOp :: UFunc -> Expr -> Expr
Unary operation for most functions (eg. sin, cos, log, etc.).

UnaryOpB :: UFuncB -> Expr -> Expr
Unary operation for Bool -> Bool operations.

UnaryOpVV :: UFuncVV -> Expr -> Expr
Unary operation for Vector -> Vector operations.

UnaryOpVN :: UFuncVN -> Expr -> Expr
Unary operation for Vector -> Number operations.

ArithBinaryOp :: ArithBinOp -> Expr -> Expr -> Expr
Binary operator for arithmetic between expressions (fractional, power, and subtraction).

BoolBinaryOp :: BoolBinOp -> Expr -> Expr -> Expr
Binary operator for boolean operators (implies, iff).

EqBinaryOp :: EqBinOp -> Expr -> Expr -> Expr
Binary operator for equality between expressions.

LABinaryOp :: LABinOp -> Expr -> Expr -> Expr
Binary operator for indexing two expressions.

OrdBinaryOp :: OrdBinOp -> Expr -> Expr -> Expr
Binary operator for ordering expressions (less than, greater than, etc.).

VVVBinaryOp :: VVVBinOp -> Expr -> Expr -> Expr
Binary operator for Vector x Vector -> Vector operations (cross product).

VVNBinaryOp :: VVNBinOp -> Expr -> Expr -> Expr
Binary operator for Vector x Vector -> Number operations (dot product).

Operator :: AssocArithOper -> DomainDesc Expr Expr -> Expr -> Expr
Operators are generalized arithmetic operators over a DomainDesc of an Expr. Could be called BigOp. ex: Summation is represented via Add over a discrete domain.

RealI :: UID -> RealInterval Expr Expr -> Expr
A different kind of IsIn. A UID is an element of an interval.

data UFunc

Unary functions (abs, log, ln, sin, etc.).

Constructors

Abs

Log

Ln

Sin

Cos

Tan

Sec

Csc

Cot

Arcsin

Arccos

Arctan

Exp

Sqrt

Neg

data UFuncB

Bool -> Bool operators.

Constructors

Not

data UFuncVV

Vector -> Vector operators.

Constructors

NegV

data UFuncVN

Vector -> Number operators.

Constructors

Norm

Dim

data ArithBinOp

Arithmetic operators (fractional, power, and subtraction).

Constructors

Frac

Pow

Subt

data BoolBinOp

Conditional and Biconditional operators (Expressions can imply one another, or exist if and only if another expression exists).

Constructors

Impl

Iff

data EqBinOp

Equality operators (equal or not equal).

Constructors

Eq

NEq

data LABinOp

Index operator.

Constructors

Index

data OrdBinOp

Ordered binary operators (less than, greater than, less than or equal to, greater than or equal to).

Constructors

Lt

Gt

LEq

GEq

data VVVBinOp

Vector x Vector -> Vector binary operations (cross product).

Constructors

Cross

data VVNBinOp

Vector x Vector -> Number binary operations (dot product).

Constructors

Dot

data AssocArithOper

Associative operators (adding and multiplication). Also specifies whether it is for integers or for real numbers.

Constructors

AddI

AddRe

MulI

MulRe

data AssocBoolOper

Associative boolean operators (and, or).

Constructors

And

Or

data DerivType

Determines the type of the derivative (either a partial derivative or a total derivative).

Constructors

Part

Total

data Completeness

For case expressions (either complete or incomplete).

Constructors

Complete

Incomplete

type Relation = Expr

A relation is just an expression (Expr).

Expr Precedence

precA :: AssocArithOper -> Int

precA - precedence for arithmetic-related Binary-Associative (Commutative) operators.

precB :: AssocBoolOper -> Int

precB - precedence for boolean-related Binary-Associative (Commutative) operators.

eprec :: Expr -> Int

eprec - Expression precedence.

mePrec :: ModelExpr -> Int

eprec - ModelExpr precedence.

Extracting UID Functions

eDep :: Expr -> [String]

Get dependencies from an equation.

eNames :: Expr -> [String]

Generic traverse of all expressions that could lead to names.

eNames' :: Expr -> [String]

Generic traverse of all positions that could lead to eNames without functions. FIXME : this should really be done via post-facto filtering, but right now the information needed to do this is not available!

eNamesRI :: RealInterval Expr Expr -> [String]

Generic traversal of everything that could come from an interval to names (similar to eNames).

meDep :: ModelExpr -> [String]

Get dependencies from an equation.

sdep :: Sentence -> [UID]

This is to collect symbolic UIDs that are printed out as a Symbol.

lnames :: Sentence -> [UID]

Generic traverse of all positions that could lead to reference UIDs from Sentences.

lnames' :: [Sentence] -> [UID]

Get reference UIDs from Sentences.