Language.Drasil.Development

Documentation

class NounPhrase n where

Minimal complete definition

phraseNP, pluralNP, sentenceCase, titleCase

Methods

phraseNP :: n -> Sentence

Retrieves singular form of term. Ex. "the quick brown fox".

pluralNP :: n -> PluralForm

Retrieves plural form of term. Ex. "the quick brown foxes".

data DisplayExpr where

A variant of the expression language that allows for extensions of Expr that wouldn't normally type check, or that we only care for displaying.

Constructors

AlgebraicExpr :: Expr -> DisplayExpr

SpaceExpr :: Space -> DisplayExpr

BinOp :: DisplayBinOp -> DisplayExpr -> DisplayExpr -> DisplayExpr

AssocBinOp :: DisplayAssocBinOp -> [DisplayExpr] -> DisplayExpr

data DisplayBinOp

Binary Display Operations.

Constructors

Defines

IsIn

data DisplayAssocBinOp

Associative Binary Display Operations.

Constructors

And

Equivalence

data UFuncB

Bool -> Bool operators.

Constructors

Not

data UFuncVV

Vector -> Vector operators.

Constructors

NegV

data UFuncVN

Vector -> Number operators.

Constructors

Norm

Dim

data ArithBinOp

Arithmetic operators (fractional, power, and subtraction).

Constructors

Frac

Pow

Subt

data BoolBinOp

Conditional and Biconditional operators (Expressions can imply one another, or exist if and only if another expression exists).

Constructors

Impl

Iff

data EqBinOp

Equality operators (equal or not equal).

Constructors

Eq

NEq

data LABinOp

Index operator.

Constructors

Index

data OrdBinOp

Ordered binary operators (less than, greater than, less than or equal to, greater than or equal to).

Constructors

Lt

Gt

LEq

GEq

data VVVBinOp

Vector x Vector -> Vector binary operations (cross product).

Constructors

Cross

data VVNBinOp

Vector x Vector -> Number binary operations (dot product).

Constructors

Dot

eDep :: Expr -> [String]

Get dependencies from an equation.

eNames :: Expr -> [String]

Generic traverse of all expressions that could lead to names.

eNames' :: Expr -> [String]

Generic traverse of all positions that could lead to eNames without functions. FIXME : this should really be done via post-facto filtering, but right now the information needed to do this is not available!

eNamesRI :: RealInterval Expr Expr -> [String]

Generic traversal of everything that could come from an interval to names (similar to eNames).

deDep :: DisplayExpr -> [String]

Get dependencies from display expressions.

sdep :: Sentence -> [UID]

This is to collect symbolic UIDs that are printed out as a Symbol.

lnames :: Sentence -> [UID]

Generic traverse of all positions that could lead to reference UIDs from Sentences.

lnames' :: [Sentence] -> [UID]

Get reference UIDs from Sentences.

precA :: AssocArithOper -> Int

precA - precedence for arithmetic-related Binary-Associative (Commutative) operators.

precB :: AssocBoolOper -> Int

precB - precedence for boolean-related Binary-Associative (Commutative) operators.

eprec :: Expr -> Int

eprec - Expression precedence.

dePrec :: DisplayExpr -> Int

dePrec - "Display Expression" precedence.

dePrecAssoc :: DisplayAssocBinOp -> Int

dePrecAssoc - precedence for associative binary operators.