Language.Drasil.Document

Description

Document Description Language.

Section Types

data SecCons

Section Contents are split into subsections or contents, where contents are standard layout objects (see Contents).

Constructors

Sub Section

Con Contents

data Partition

Constructors

Sections

Part

Chapter

data Section

Sections have a title (Sentence), a list of contents (SecCons) and a shortname (Reference).

Constructors

Section

Fields

tle :: Title

cons :: [SecCons]

_lab :: Reference

lab :: Lens' Section Reference

data Document

A Document has a Title (Sentence), Author(s) (Sentence), and Sections which hold the contents of the document.

Constructors

Document Title Author ShowTableOfContents [Section]

Notebook Title Author [Section]

data ShowTableOfContents

Determines whether or not the table of contents appears on the generated artifacts.

Constructors

ToC

NoToC

checkToC :: Document -> Document

Manually removes the first section of a document (table of contents section). temp fix for Notebook (see if we need this in notebook later)

Content Constructors

llcc :: Reference -> RawContent -> LabelledContent

Smart constructor for labelled content chunks.

ulcc :: RawContent -> UnlabelledContent

Smart constructor for unlabelled content chunks (no Reference).

mkParagraph :: Sentence -> Contents

Smart constructor that wraps UnlabelledContent into Contents.

mkFig :: Reference -> RawContent -> Contents

Smart constructor that wraps LabelledContent into Contents.

mkRawLC :: RawContent -> Reference -> LabelledContent

Smart constructor similar to llcc, but takes in RawContent first.

Section Constructors

section :: Sentence -> [Contents] -> [Section] -> Reference -> Section

Smart constructor for creating Sections with a title (Sentence), introductory contents (ie. paragraphs, tables, etc.), a list of subsections, and a shortname (Reference).

fig :: Lbl -> Filepath -> RawContent

Figure smart constructor with a Lbl and a Filepath. Assumes 100% of page width as max width. Defaults to WithCaption.

figNoCap :: Lbl -> Filepath -> RawContent

Figure smart constructor without a caption.

figWithWidth :: Lbl -> Filepath -> MaxWidthPercent -> RawContent

Figure smart constructor that allows for customized max widths. Defaults to WithCaption.

figNoCapWithWidth :: Lbl -> Filepath -> MaxWidthPercent -> RawContent

Figure smart constructor with customized max widths and no caption.

Reference Constructors

docNs :: UID -> UID

docUid :: String -> UID

makeTabRef :: String -> Reference

Create a reference for a table. Takes in the name of a table (which will also be used for its shortname).

makeFigRef :: String -> Reference

Create a reference for a figure. Takes in the name of a figure (which will also be used for its shortname).

makeSecRef :: String -> Sentence -> Reference

Create a reference for a section. Takes in the name of a section and a shortname for the section.

makeEqnRef :: String -> Reference

Create a reference for a equation. Takes in the name of the equation (which will also be used for its shortname).

makeURI :: String -> String -> ShortName -> Reference

Create a reference for a URI. Takes in a UID (as a String), a reference address, and a shortname.

makeTabRef' :: UID -> Reference

Variants of makeTabRef that takes a UID instead of a String.

makeFigRef' :: UID -> Reference

Variants of makeFigRef that takes a UID instead of a String.

makeSecRef' :: UID -> Sentence -> Reference

Variants of makeSecRef that takes a UID instead of a String.

makeEqnRef' :: UID -> Reference

Variants of makeEqnRef that takes a UID instead of a String.

makeURI' :: UID -> String -> ShortName -> Reference

Variants of makeURI that takes a UID instead of a String.