Language.Drasil.Label.Type

Description

Even though we do not have Labels per se, here we define the different ways of construction ways to mark labels.

Types

data LblType

Applying different pieces of information for a reference. An RP is a decorated internal reference. Citation is a citation. URI is for URLs and other external links.

Constructors

RP IRefProg String

Citation String

URI String

data IRefProg

Created for different forms of references. Used in LblType.

Constructors

Deferred UID
Deferred lookup; done later. Used for domains in a ConceptInstance.

RS String
Lifts a String into a RefProg.

RConcat IRefProg IRefProg
Concatenates with two subprograms.

Name
The Symbol to insert the ShortName directly.

Classes

class HasRefAddress b where

Members must have a reference address.

Methods

getRefAdd :: b -> LblType

Provides the ability to hold a reference address.

class (HasUID s, HasRefAddress s) => Referable s where

Members of this class have the ability to be referenced.

Methods

refAdd :: s -> String

The referencing address (what we're linking to). Only visible in the source (tex/html).

renderRef :: s -> LblType

Alternate form of reference.

LblType accessor

getAdd :: LblType -> String

Retrieves the String contained in a LblType.

IRefProg constructors

name :: IRefProg

Constructor that gets the Name of an IRefProg.

(+::+) :: IRefProg -> IRefProg -> IRefProg

Constructor that concatenates two subprograms.

raw :: String -> IRefProg

Constructor for a String into an IRefProg.

defer :: UID -> IRefProg

Constructor to defer a UID lookup; done later.

prepend :: String -> IRefProg

Prepends a String to an IRefProg.