Language.Drasil.Reference

Description

Defines a type used to hold referencing information.

Type

data Reference

A Reference contains the identifier (UID), a reference address (LblType), a human-readable shortname (ShortName), and any extra information about the reference (RefInfo).

Constructors

Reference UID LblType ShortName

Class

class HasReference c where

A class that contains a list of References.

Methods

getReferences :: Lens' c [Reference]

Provides a Lens to the References.

Constructors

ref :: (HasUID r, HasRefAddress r, HasShortName r) => r -> Reference

Projector function that creates a Reference from something Referable.

refS :: (HasUID r, HasRefAddress r, HasShortName r) => r -> Sentence

Takes the reference UID and wraps it into a Sentence.

namedRef :: (HasUID r, HasRefAddress r, HasShortName r) => r -> Sentence -> Sentence

Takes a Reference with a name to be displayed and wraps it into a Sentence. Does not overwrite the shortname contained in the reference, but will only display as the given Sentence.

complexRef :: (HasUID r, HasRefAddress r, HasShortName r) => r -> RefInfo -> Sentence

Takes a Reference with additional display info. Uses the internal shortname for its display name.

namedComplexRef :: (HasUID r, HasRefAddress r, HasShortName r) => r -> Sentence -> RefInfo -> Sentence

Takes a Reference with a name to be displayed and any additional information and wraps it into a Sentence. Does not overwrite the shortname contained in the reference, but will only display as the given Sentence along with the given RefInfo.