Language.Drasil.Space

Description

Number space types and functions.

Types

data Space

The difference kinds of spaces that may exist. This type holds numerical spaces (such as the set of integers, rationals, etc.), a space for booleans, a space for characters, dimensional spaces (vectors, arrays, etc.), a space for Actors, discrete sets (both for numbers and strings), and a void space.

Constructors

Integer

Rational

Real

Natural

Boolean

Char

String

Vect Space

Set Space

Matrix Int Int Space

Array Space

Actor String

Function (NonEmpty Primitive) Primitive

Void

type Primitive = Space

data RealInterval a b where

A RealInterval is a subset of Space (as a Space). These come in different flavours. For now, we embed Expr for the bounds, but that will change as well.

Constructors

Bounded

Fields

:: (Inclusive, a)

-> (Inclusive, b)

-> RealInterval a b

Interval from (x .. y).

UpTo

Fields

:: (Inclusive, a)

-> RealInterval a b

Interval from (-infinity .. x).

UpFrom

Fields

:: (Inclusive, b)

-> RealInterval a b

Interval from (x .. infinity).

data Inclusive

Inclusive or exclusive bounds.

Constructors

Inc

Exc

data DomainDesc (tplgy :: RTopology) a b where

Describes the domain of a Symbol given a topology. Can be bounded or encase all of the domain.

Constructors

BoundedDD :: Symbol -> RTopology -> a -> b -> DomainDesc 'Discrete a b

AllDD :: Symbol -> RTopology -> DomainDesc 'Continuous a b

data RTopology

Topology of a subset of reals.

Constructors

Continuous

Discrete

type DiscreteDomainDesc a b = DomainDesc 'Discrete a b

type ContinuousDomainDesc a b = DomainDesc 'Continuous a b

Class

class HasSpace c where

HasSpace is anything which has a Space.

Methods

typ :: Getter c Space

Provides a Getter to the Space.

Functions

getActorName :: Space -> String

Gets the name of an Actor.

getInnerSpace :: Space -> Space

Gets the inner Space of a vector or set.

mkFunction :: [Primitive] -> Primitive -> Space

isBasicNumSpace :: Space -> Bool

Is this Space a basic numeric space?

assertReal :: Space -> (String -> String) -> Either String ()

Assert that a Space is Space or return a formatted error message.

assertNumeric :: Space -> (String -> String) -> Either String ()

Assert that a Space is "numeric" or return a formatted error message.

assertNonNatNumeric :: Space -> (String -> String) -> Either String ()

Assert that a numeric Space is not Natural or return a formatted error message.

assertEquivNumeric :: Space -> Space -> (String -> String -> String) -> Either String ()

Assert that two numeric Spaces are equivalent or return a formatted error message.

assertIndexLike :: Space -> (String -> String) -> Either String ()

Assert that a Space is an index-like type (Integer or Natural) or return a formatted error message.

assertBoolean :: Space -> (String -> String) -> Either String ()

Assert that a Space is a Boolean or return a formatted error message.

assertSet :: Space -> (String -> String) -> Either String Space

Assert that a Space is a Set and return either the element type or a formatted error message.

assertVector :: Space -> (String -> String) -> Either String Space

Assert that a Space is a Vect and return either the element type or a formatted error message.

assertNumericVector :: Space -> (String -> String) -> Either String Space

Assert that a Space is a numeric vector (i.e., a vector of a basic numeric type) and return either the numeric type or a formatted error message.

assertNonNatNumVector :: (String -> String) -> Space -> Either String Space

Assert that a Space is a non-Natural numeric vector and return either the numeric type or a formatted error message.

assertRealVector :: Space -> (String -> String) -> Either String ()

Assert that a Space is a Vect of Spaces or return a formatted error message.

assertFunction :: Space -> (String -> String) -> Either String (NonEmpty Primitive, Primitive)

Assert that a Space is a Function, returning either the parameters and output type or a formatted error message.

assertNonFunction :: Space -> (String -> String) -> Either String ()

Assert that a Space is anything but a Function or return a formatted error message.