Language.Drasil.Uncertainty

Description

Defines uncertainty types and functions.

Type

data Uncertainty

Something that may contain an uncertainty value and a precision value.

Class

class HasUncertainty c where

HasUncertainty is just a chunk with some uncertainty associated to it. This uncertainty is represented as a decimal value between 0 and 1 (percentage).

Methods

unc :: Lens' c Uncertainty

Provides the Lens to an Uncertainty.

Lenses

uncert :: Lens' Uncertainty (Maybe Double)

prec :: Lens' Uncertainty (Maybe Int)

Constructors

uncty :: Double -> Maybe Int -> Uncertainty

Smart constructor for values with uncertainty.

exact :: Uncertainty

Smart constructor for exact values (no uncertainty).

Constructor

defaultUncrt :: Uncertainty

The default uncertainty is set to 0.1.

Accessors

uncVal :: HasUncertainty x => x -> Double

Accessor for uncertainty value from something that has an uncertainty.

uncPrec :: HasUncertainty x => x -> Maybe Int

Accessor for precision value from something that has an uncertainty.